An S4 class to Represent a Dose Rate Calibration Curve

An S4 class to Represent a Dose Rate Calibration Curve

Slots

Ni

A DoseRateModel object.

NiEi

A DoseRateModel object.

details

A list of length-one vector giving the curve metadata.

slope

A numeric vector.

intercept

A numeric vector.

covariance

A numeric vector.

MSWD

A numeric vector.

df

A numeric vector.

p_value

A numeric vector.

data

A data.frame.

range

A numeric vector.

background

A numeric vector.

Subset

In the code snippets below, x is a CalibrationCurve object.

x[[i]]

Extracts information from a slot selected by subscript i. i is a character vector of length one.

See also

Author

N. Frerebeau