An S4 class to represent a gamma sectrum

Represents a single spectrum of a gamma ray spectrometry measurement.

# S4 method for GammaSpectrum
[[(x, i)

# S4 method for GammaSpectrum
length(x)

Arguments

x
An object of class GammaSpectrum.

i
A length-one character vector specifying the element to extract or replace (see below). Character sring will be matched to the names of the slots.

Slots

hash

A character string giving the 32-byte MD5 hash of the imported file.

reference

A character string the measurement reference.

date

A POSIXct element giving the measurement date and time.

instrument

A character string giving the instrument name.

file_format

A character string giving the format of the imported file.

live_time

A numeric value.

real_time

A numeric value.

chanel

A integer vector giving the channel number. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero).

energy

A numeric vector giving the gamma ray's energy (in keV).

counts

A integer vector giving the counts number for each channel. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero).

rate

A numeric vector the count rate (in 1/s) for each channel.

calibration

A linear model used for energy scale calibration (see calibrate).

Methods

estimateBaseline

Estimate the baseline of a GammaSpectrum object. See estimateBaseline for details.

predict

Estimate the in-situ gamma dose rate of a GammaSpectrum object. See predict for details.

findPeaks

Look for local maxima to extract peaks out of a GammaSpectrum object. See findPeaks for details.

fitPeaks

Adjust a Gaussian at given positions to estimate peak paramaters out of a GammaSpectrum object. See fitPeaks for details.

removeBaseline

Estimate and remove the baseline of a GammaSpectrum object. See removeBaseline for details.

Access

In the code snippets below, x is a GammaSpectrum object.

length(x)

Get the number of chanel in x.

Coerce

In the code snippets below, x is a GammaSpectrum object.

as(x, "matrix")

Coerces x to a matrix.

as(x, "data.frame")

Coerces x to a data frame.

Subset

In the code snippets below, x is a GammaSpectrum object.

x[[i]]

Extracts informations from a slot selected by subscript i. i is a character vector of length one.

See also

Examples

# Import a Canberra CNF file
spc_file <- system.file("extdata/test.cnf", package = "gamma")
(spectrum <- read(spc_file))

#> Gamma spectrum:
#>   Reference: test
#>   Instrument: InSpector 1000
#>   Date: 2019-02-07 11:48:18
#>   Number of chanels: 1024
#>   Energy range (keV): -7-3124.91

# Coerce
df <- as(spectrum, "data.frame")
head(df)

#>   chanel    energy counts rate
#> 1      1 -7.004032      0    0
#> 2      2 -4.007936      0    0
#> 3      3 -1.011712      0    0
#> 4      4  1.984640      0    0
#> 5      5  4.981120      0    0
#> 6      6  7.977728      0    0

tail(df)

#>      chanel   energy counts         rate
#> 1019   1019 3109.281      1 0.0002953739
#> 1020   1020 3112.408      1 0.0002953739
#> 1021   1021 3115.534      2 0.0005907477
#> 1022   1022 3118.661      0 0.0000000000
#> 1023   1023 3121.788      1 0.0002953739
#> 1024   1024 3124.915      0 0.0000000000

# Subset
spectrum[["hash"]]

#> [1] "e0a2c67173cf7f407db7148ae0058bbf"

spectrum[["reference"]]

#> [1] "test"

spectrum[["date"]]

#> [1] "2019-02-07 11:48:18 CET"

spectrum[["instrument"]]

#> [1] "InSpector 1000"

spectrum[["file_format"]]

#> [1] "Canberra CNF"

spectrum[["live_time"]]

#> [1] 3385.54

spectrum[["real_time"]]

#> [1] 3403.67

spectrum[["calibration"]]

#> NULL