An S4 Class to Represent a Gamma Sectrum

Represents a single spectrum of a gamma ray spectrometry measurement.

Slots

hash

A character string giving the 32-byte MD5 hash of the imported file.

name

A character string the measurement reference.

date

A POSIXct element giving the measurement date and time.

instrument

A character string giving the instrument name.

file_format

A character string giving the format of the imported file.

live_time

A numeric value.

real_time

A numeric value.

chanel

A integer vector giving the channel number. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero).

energy

A numeric vector giving the gamma ray's energy (in keV).

count

A numeric vector giving the counts number for each channel. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero).

rate

A numeric vector the count rate (in 1/s) for each channel.

calibration

A linear model used for energy scale calibration (see calibrate_energy).

dose_rate

A length-two numeric vector giving the dose rate and corresponding error.

Note

This class retains copy construction.

Access

In the code snippets below, x is a GammaSpectrum object.

get_hash(x)

Get the MD5 hash of the raw data file.

get_names(x), set_names(x) <- value

Retrieves or sets the name of x according to value.

get_chanels(x)

Get the number of chanels in x.

get_energy(x)

Get the energy range of x.

get_dose(x)

Get the dose rate of x.

Coerce

In the code snippets below, x is a GammaSpectrum object.

as(x, "data.frame")

Coerces x to a data frame.

Subset

In the code snippets below, x is a GammaSpectrum object.

x[[i]]

Extracts informations from a slot selected by subscript i. i is a character vector of length one and will be matched to the name of the slots.

See also

Examples

# Import a Canberra CNF file
spc_file <- system.file("extdata/test_CNF.cnf", package = "gamma")
(spectrum <- read(spc_file))

#> Gamma spectrum:
#> *  name: test_CNF
#> *  date: 2019-02-07 11:48:18
#> *  live_time: 3385.54
#> *  real_time: 3403.67
#> *  chanels: 1024
#> *  energy_min: -7.00403203118913
#> *  energy_max: 3124.91452789307

## Access
get_hash(spectrum)

#> [1] "e0a2c67173cf7f407db7148ae0058bbf"

get_names(spectrum)

#> [1] "test_CNF"

get_chanels(spectrum)

#> [1] 1024

get_energy(spectrum)

#> [1]   -7.004032 3124.914528

get_dose(spectrum)

#> [1] 0 0

# Coerce
as(spectrum, "data.frame")

#>      chanel      energy count         rate
#> 1         1   -7.004032     0 0.0000000000
#> 2         2   -4.007936     0 0.0000000000
#> 3         3   -1.011712     0 0.0000000000
#> 4         4    1.984640     0 0.0000000000
#> 5         5    4.981120     0 0.0000000000
#> 6         6    7.977728     0 0.0000000000
#> 7         7   10.974464     0 0.0000000000
#> 8         8   13.971328     0 0.0000000000
#> 9         9   16.968320     0 0.0000000000
#> 10       10   19.965440     0 0.0000000000
#> 11       11   22.962688     0 0.0000000000
#> 12       12   25.960064     0 0.0000000000
#> 13       13   28.957568  1790 0.5287191999
#> 14       14   31.955200  7194 2.1249195106
#> 15       15   34.952960  7220 2.1325992308
#> 16       16   37.950848  7167 2.1169444165
#> 17       17   40.948863  7425 2.1931508711
#> 18       18   43.947007  8277 2.4448093953
#> 19       19   46.945279  9078 2.6814038529
#> 20       20   49.943679  9813 2.8985036361
#> 21       21   52.942207 11231 3.3173437620
#> 22       22   55.940863 13055 3.8561056730
#> 23       23   58.939647 15211 4.4929317037
#> 24       24   61.938559 17220 5.0863377777
#> 25       25   64.937599 19721 5.8250677883
#> 26       26   67.936767 21336 6.3020965636
#> 27       27   70.936063 23656 6.9873639065
#> 28       28   73.935487 26034 7.6897629329
#> 29       29   76.935039 29035 8.5761798709
#> 30       30   79.934719 30838 9.1087389309
#> 31       31   82.934527 31276 9.2381126792
#> 32       32   85.934463 31239 9.2271838466
#> 33       33   88.934527 30668 9.0585253756
#> 34       34   91.934719 31299 9.2449062779
#> 35       35   94.935039 31337 9.2561304844
#> 36       36   97.935487 30825 9.1048990708
#> 37       37  100.936063 30081 8.8851409229
#> 38       38  103.936767 29632 8.7525180621
#> 39       39  106.937599 29312 8.6579984286
#> 40       40  109.938559 29092 8.5930161806
#> 41       41  112.939647 28785 8.5023364072
#> 42       42  115.940863 28481 8.4125427554
#> 43       43  118.942207 28049 8.2849412501
#> 44       44  121.943679 28140 8.3118202709
#> 45       45  124.945279 27339 8.0752258133
#> 46       46  127.947007 27048 7.9892720216
#> 47       47  130.948863 26291 7.7656740136
#> 48       48  133.950847 26266 7.7582896672
#> 49       49  136.952958 25809 7.6233038156
#> 50       50  139.955198 25329 7.4815243654
#> 51       51  142.957566 24432 7.2165740177
#> 52       52  145.960062 24064 7.1078764392
#> 53       53  148.962686 23427 6.9197232938
#> 54       54  151.965438 22911 6.7673103848
#> 55       55  154.968318 22382 6.6110576156
#> 56       56  157.971326 21511 6.3537869882
#> 57       57  160.974462 21168 6.2524737560
#> 58       58  163.977726 20857 6.1606124872
#> 59       59  166.981118 20073 5.9290393851
#> 60       60  169.984638 19566 5.7792848408
#> 61       61  172.988286 18943 5.5952669294
#> 62       62  175.992062 18429 5.4434447680
#> 63       63  178.995966 18224 5.3828931278
#> 64       64  181.999998 17632 5.2080318059
#> 65       65  185.004158 17607 5.2006474595
#> 66       66  188.008446 16901 4.9921135181
#> 67       67  191.012862 16812 4.9658252450
#> 68       68  194.017406 16350 4.8293625241
#> 69       69  197.022078 16236 4.7956899047
#> 70       70  200.026878 15639 4.6193517135
#> 71       71  203.031806 15406 4.5505296053
#> 72       72  206.036862 14973 4.4226327262
#> 73       73  209.042046 14355 4.2400916840
#> 74       74  212.047358 14223 4.2011023352
#> 75       75  215.052798 13909 4.1083549449
#> 76       76  218.058366 13629 4.0256502655
#> 77       77  221.064062 13191 3.8962765172
#> 78       78  224.069886 13021 3.8460629619
#> 79       79  227.075838 13120 3.8753049735
#> 80       80  230.081918 12957 3.8271590352
#> 81       81  233.088125 12663 3.7403191219
#> 82       82  236.094461 12926 3.8180024457
#> 83       83  239.100925 13103 3.8702836180
#> 84       84  242.107517 13260 3.9166573132
#> 85       85  245.114237 13286 3.9243370334
#> 86       86  248.121085 13135 3.8797355813
#> 87       87  251.128061 12951 3.8253867921
#> 88       88  254.135165 12127 3.5819987358
#> 89       89  257.142397 11400 3.3672619434
#> 90       90  260.149757 10681 3.1548881419
#> 91       91  263.157245  9803 2.8955498975
#> 92       92  266.164861  9336 2.7576103074
#> 93       93  269.172605  8947 2.6427098779
#> 94       94  272.180477  8567 2.5304678131
#> 95       95  275.188477  8327 2.4595780880
#> 96       96  278.196605  8196 2.4208841130
#> 97       97  281.204861  8020 2.3688983146
#> 98       98  284.213245  7952 2.3488128925
#> 99       99  287.221757  7754 2.2903288692
#> 100     100  290.230397  7456 2.2023074606
#> 101     101  293.239165  7405 2.1872433940
#> 102     102  296.248061  7307 2.1582967562
#> 103     103  299.257085  7245 2.1399835772
#> 104     104  302.266237  7384 2.1810405430
#> 105     105  305.275517  7101 2.0974497421
#> 106     106  308.284925  6806 2.0103144550
#> 107     107  311.294461  6637 1.9603962736
#> 108     108  314.304125  6439 1.9019122503
#> 109     109  317.313917  6294 1.8590830414
#> 110     110  320.323837  6030 1.7811043438
#> 111     111  323.333885  5774 1.7054886370
#> 112     112  326.344061  5566 1.6440508752
#> 113     113  329.354364  5577 1.6472999876
#> 114     114  332.364796  5486 1.6204209668
#> 115     115  335.375356  5320 1.5713889069
#> 116     116  338.386044  5292 1.5631184390
#> 117     117  341.396860  5339 1.5770010102
#> 118     118  344.407804  5426 1.6026985355
#> 119     119  347.418876  5495 1.6230793315
#> 120     120  350.430076  5530 1.6334174164
#> 121     121  353.441404  5602 1.6546843340
#> 122     122  356.452860  5357 1.5823177396
#> 123     123  359.464444  5372 1.5867483474
#> 124     124  362.476156  5327 1.5734565239
#> 125     125  365.487996  5160 1.5241290902
#> 126     126  368.499964  5016 1.4815952551
#> 127     127  371.512060  4887 1.4434920279
#> 128     128  374.524284  4667 1.3785097798
#> 129     129  377.536636  4325 1.2774919215
#> 130     130  380.549116  4219 1.2461822929
#> 131     131  383.561724  3937 1.1628868659
#> 132     132  386.574460  3642 1.0757515788
#> 133     133  389.587324  3582 1.0580291475
#> 134     134  392.600316  3289 0.9714846081
#> 135     135  395.613436  3324 0.9818226930
#> 136     136  398.626684  3203 0.9460824566
#> 137     137  401.640060  3174 0.9375166148
#> 138     138  404.653564  3142 0.9280646514
#> 139     139  407.667196  2982 0.8808048347
#> 140     140  410.680956  3046 0.8997087614
#> 141     141  413.694844  2937 0.8675130112
#> 142     142  416.708860  2931 0.8657407681
#> 143     143  419.723004  2869 0.8474275891
#> 144     144  422.737276  2882 0.8512674492
#> 145     145  425.751675  2895 0.8551073093
#> 146     146  428.766203  2894 0.8548119355
#> 147     147  431.780859  2680 0.7916019306
#> 148     148  434.795643  2604 0.7691535176
#> 149     149  437.810555  2597 0.7670859006
#> 150     150  440.825595  2572 0.7597015543
#> 151     151  443.840763  2592 0.7656090314
#> 152     152  446.856059  2607 0.7700396392
#> 153     153  449.871483  2517 0.7434559922
#> 154     154  452.887035  2568 0.7585200588
#> 155     155  455.902715  2480 0.7325271596
#> 156     156  458.918523  2426 0.7165769715
#> 157     157  461.934459  2449 0.7233705701
#> 158     158  464.950523  2298 0.6787691181
#> 159     159  467.966715  2394 0.7071250081
#> 160     160  470.983035  2367 0.6991499140
#> 161     161  473.999483  2329 0.6879257076
#> 162     162  477.016059  2259 0.6672495377
#> 163     163  480.032763  2267 0.6696125286
#> 164     164  483.049595  2290 0.6764061272
#> 165     165  486.066555  2291 0.6767015011
#> 166     166  489.083643  2224 0.6569114528
#> 167     167  492.100859  2041 0.6028580374
#> 168     168  495.118203  2219 0.6554345836
#> 169     169  498.135675  2062 0.6090608884
#> 170     170  501.153275  2076 0.6131961223
#> 171     171  504.171003  2011 0.5939968218
#> 172     172  507.188859  2043 0.6034487851
#> 173     173  510.206843  2088 0.6167406086
#> 174     174  513.224955  2113 0.6241249550
#> 175     175  516.243195  1960 0.5789327552
#> 176     176  519.261563  1989 0.5874985970
#> 177     177  522.280058  2079 0.6140822439
#> 178     178  525.298682  2131 0.6294416843
#> 179     179  528.317434  2079 0.6140822439
#> 180     180  531.336314  2064 0.6096516361
#> 181     181  534.355322  2037 0.6016765420
#> 182     182  537.374458  2068 0.6108331315
#> 183     183  540.393722  1884 0.5564843422
#> 184     184  543.413114  1834 0.5417156495
#> 185     185  546.432634  1926 0.5688900441
#> 186     186  549.452282  1858 0.5488046220
#> 187     187  552.472058  1795 0.5301960692
#> 188     188  555.491962  1735 0.5124736379
#> 189     189  558.511994  1675 0.4947512066
#> 190     190  561.532154  1615 0.4770287753
#> 191     191  564.552442  1735 0.5124736379
#> 192     192  567.572858  1688 0.4985910667
#> 193     193  570.593402  1643 0.4852992433
#> 194     194  573.614074  1712 0.5056800392
#> 195     195  576.634874  1689 0.4988864406
#> 196     196  579.655802  1759 0.5195626104
#> 197     197  582.676858  1810 0.5346266770
#> 198     198  585.698042  1836 0.5423063972
#> 199     199  588.719354  1937 0.5721391565
#> 200     200  591.740794  2002 0.5913384571
#> 201     201  594.762362  1964 0.5801142506
#> 202     202  597.784058  2020 0.5966551865
#> 203     203  600.805882  2103 0.6211712164
#> 204     204  603.827834  2105 0.6217619641
#> 205     205  606.849914  2173 0.6418473862
#> 206     206  609.872122  2047 0.6046302805
#> 207     207  612.894458  2186 0.6456872463
#> 208     208  615.916922  2118 0.6256018242
#> 209     209  618.939513  2116 0.6250110765
#> 210     210  621.962233  2163 0.6388936477
#> 211     211  624.985081  2076 0.6131961223
#> 212     212  628.008057  2099 0.6199897210
#> 213     213  631.031161  2011 0.5939968218
#> 214     214  634.054393  1884 0.5564843422
#> 215     215  637.077753  1731 0.5112921425
#> 216     216  640.101241  1719 0.5077476562
#> 217     217  643.124857  1565 0.4622600826
#> 218     218  646.148601  1533 0.4528081192
#> 219     219  649.172473  1498 0.4424700343
#> 220     220  652.196473  1374 0.4058436763
#> 221     221  655.220601  1319 0.3895981143
#> 222     222  658.244857  1282 0.3786692817
#> 223     223  661.269241  1215 0.3588792334
#> 224     224  664.293753  1193 0.3523810086
#> 225     225  667.318393  1155 0.3411568022
#> 226     226  670.343161  1117 0.3299325957
#> 227     227  673.368057  1139 0.3364308205
#> 228     228  676.393081  1040 0.3071888089
#> 229     229  679.418233  1131 0.3340678297
#> 230     230  682.443513  1075 0.3175268938
#> 231     231  685.468921  1048 0.3095517997
#> 232     232  688.494457  1024 0.3024628272
#> 233     233  691.520121  1048 0.3095517997
#> 234     234  694.545913  1074 0.3172315199
#> 235     235  697.571833  1049 0.3098471736
#> 236     236  700.597881  1079 0.3187083892
#> 237     237  703.624057  1079 0.3187083892
#> 238     238  706.650361  1001 0.2956692285
#> 239     239  709.676793  1006 0.2971460978
#> 240     240  712.703353  1045 0.3086656781
#> 241     241  715.730040  1000 0.2953738547
#> 242     242  718.756856   997 0.2944877331
#> 243     243  721.783800  1035 0.3057119396
#> 244     244  724.810872  1030 0.3042350703
#> 245     245  727.838072  1024 0.3024628272
#> 246     246  730.865400   942 0.2782421711
#> 247     247  733.892856  1014 0.2995090887
#> 248     248  736.920440  1036 0.3060073135
#> 249     249  739.948152   980 0.2894663776
#> 250     250  742.975992  1070 0.3160500245
#> 251     251  746.003960   966 0.2853311436
#> 252     252  749.032056  1084 0.3201852585
#> 253     253  752.060280  1001 0.2956692285
#> 254     254  755.088632   923 0.2726300679
#> 255     255  758.117112   983 0.2903524992
#> 256     256  761.145720   982 0.2900571253
#> 257     257  764.174456   946 0.2794236665
#> 258     258  767.203320   912 0.2693809555
#> 259     259  770.232312  1007 0.2974414717
#> 260     260  773.261432   966 0.2853311436
#> 261     261  776.290680   950 0.2806051620
#> 262     262  779.320056   951 0.2809005358
#> 263     263  782.349560   947 0.2797190404
#> 264     264  785.379192   936 0.2764699280
#> 265     265  788.408952   924 0.2729254417
#> 266     266  791.438840   916 0.2705624509
#> 267     267  794.468856   893 0.2637688522
#> 268     268  797.499000   914 0.2699717032
#> 269     269  800.529272   923 0.2726300679
#> 270     270  803.559672   859 0.2537261412
#> 271     271  806.590200   919 0.2714485725
#> 272     272  809.620856   898 0.2652457215
#> 273     273  812.651639   912 0.2693809555
#> 274     274  815.682551   892 0.2634734784
#> 275     275  818.713591   885 0.2614058614
#> 276     276  821.744759   804 0.2374805792
#> 277     277  824.776055   877 0.2590428706
#> 278     278  827.807479   821 0.2425019347
#> 279     279  830.839031   799 0.2360037099
#> 280     280  833.870711   820 0.2422065608
#> 281     281  836.902519   757 0.2235980080
#> 282     282  839.934455   823 0.2430926824
#> 283     283  842.966519   795 0.2348222145
#> 284     284  845.998711   770 0.2274378681
#> 285     285  849.031031   809 0.2389574484
#> 286     286  852.063479   797 0.2354129622
#> 287     287  855.096055   742 0.2191674002
#> 288     288  858.128759   778 0.2298008589
#> 289     289  861.161591   759 0.2241887557
#> 290     290  864.194551   734 0.2168044093
#> 291     291  867.227639   742 0.2191674002
#> 292     292  870.260855   761 0.2247795034
#> 293     293  873.294199   720 0.2126691754
#> 294     294  876.327671   770 0.2274378681
#> 295     295  879.361271   715 0.2111923061
#> 296     296  882.394999   723 0.2135552969
#> 297     297  885.428855   772 0.2280286158
#> 298     298  888.462839   691 0.2041033336
#> 299     299  891.496951   757 0.2235980080
#> 300     300  894.531191   782 0.2309823544
#> 301     301  897.565559   794 0.2345268406
#> 302     302  900.600055   805 0.2377759530
#> 303     303  903.634679   748 0.2209396433
#> 304     304  906.669431   788 0.2327545975
#> 305     305  909.704310   836 0.2469325425
#> 306     306  912.739318   853 0.2519538980
#> 307     307  915.774454   810 0.2392528223
#> 308     308  918.809718   850 0.2510677765
#> 309     309  921.845110   868 0.2563845059
#> 310     310  924.880630   874 0.2581567490
#> 311     311  927.916278   853 0.2519538980
#> 312     312  930.952054   921 0.2720393202
#> 313     313  933.987958   872 0.2575660013
#> 314     314  937.023990   888 0.2622919830
#> 315     315  940.060150   897 0.2649503477
#> 316     316  943.096438   923 0.2726300679
#> 317     317  946.132854   881 0.2602243660
#> 318     318  949.169398   874 0.2581567490
#> 319     319  952.206070   807 0.2383667007
#> 320     320  955.242870   782 0.2309823544
#> 321     321  958.279798   868 0.2563845059
#> 322     322  961.316854   815 0.2407296916
#> 323     323  964.354038   857 0.2531353935
#> 324     324  967.391350   796 0.2351175883
#> 325     325  970.428790   804 0.2374805792
#> 326     326  973.466358   780 0.2303916067
#> 327     327  976.504054   794 0.2345268406
#> 328     328  979.541878   807 0.2383667007
#> 329     329  982.579830   812 0.2398435700
#> 330     330  985.617910   772 0.2280286158
#> 331     331  988.656118   832 0.2457510471
#> 332     332  991.694454   811 0.2395481962
#> 333     333  994.732918   694 0.2049894552
#> 334     334  997.771510   761 0.2247795034
#> 335     335 1000.810230   718 0.2120784277
#> 336     336 1003.849078   744 0.2197581479
#> 337     337 1006.888053   686 0.2026264643
#> 338     338 1009.927157   648 0.1914022578
#> 339     339 1012.966389   663 0.1958328657
#> 340     340 1016.005749   642 0.1896300147
#> 341     341 1019.045237   663 0.1958328657
#> 342     342 1022.084853   629 0.1857901546
#> 343     343 1025.124597   622 0.1837225376
#> 344     344 1028.164469   592 0.1748613220
#> 345     345 1031.204469   549 0.1621602462
#> 346     346 1034.244597   528 0.1559573953
#> 347     347 1037.284853   599 0.1769289390
#> 348     348 1040.325237   575 0.1698399664
#> 349     349 1043.365749   526 0.1553666476
#> 350     350 1046.406389   520 0.1535944044
#> 351     351 1049.447157   500 0.1476869273
#> 352     352 1052.488053   525 0.1550712737
#> 353     353 1055.529077   555 0.1639324894
#> 354     354 1058.570229   470 0.1388257117
#> 355     355 1061.611509   515 0.1521175352
#> 356     356 1064.652917   521 0.1538897783
#> 357     357 1067.694453   521 0.1538897783
#> 358     358 1070.736117   496 0.1465054319
#> 359     359 1073.777909   494 0.1459146842
#> 360     360 1076.819829   513 0.1515267875
#> 361     361 1079.861877   478 0.1411887025
#> 362     362 1082.904053   534 0.1577296384
#> 363     363 1085.946357   513 0.1515267875
#> 364     364 1088.988789   524 0.1547758999
#> 365     365 1092.031349   508 0.1500499182
#> 366     366 1095.074037   530 0.1565481430
#> 367     367 1098.116853   497 0.1468008058
#> 368     368 1101.159797   512 0.1512314136
#> 369     369 1104.202868   563 0.1662954802
#> 370     370 1107.246068   564 0.1665908540
#> 371     371 1110.289396   529 0.1562527691
#> 372     372 1113.332852   519 0.1532990306
#> 373     373 1116.376436   553 0.1633417416
#> 374     374 1119.420148   549 0.1621602462
#> 375     375 1122.463988   595 0.1757474435
#> 376     376 1125.507956   586 0.1730890788
#> 377     377 1128.552052   591 0.1745659481
#> 378     378 1131.596276   567 0.1674769756
#> 379     379 1134.640628   640 0.1890392670
#> 380     380 1137.685108   606 0.1789965559
#> 381     381 1140.729716   642 0.1896300147
#> 382     382 1143.774452   596 0.1760428174
#> 383     383 1146.819316   591 0.1745659481
#> 384     384 1149.864308   588 0.1736798266
#> 385     385 1152.909428   610 0.1801780514
#> 386     386 1155.954676   552 0.1630463678
#> 387     387 1159.000052   599 0.1769289390
#> 388     388 1162.045556   574 0.1695445926
#> 389     389 1165.091188   592 0.1748613220
#> 390     390 1168.136948   548 0.1618648724
#> 391     391 1171.182836   549 0.1621602462
#> 392     392 1174.228852   542 0.1600926292
#> 393     393 1177.274996   558 0.1648186109
#> 394     394 1180.321268   504 0.1488684228
#> 395     395 1183.367668   535 0.1580250123
#> 396     396 1186.414196   519 0.1532990306
#> 397     397 1189.460852   526 0.1553666476
#> 398     398 1192.507636   501 0.1479823012
#> 399     399 1195.554548   517 0.1527082829
#> 400     400 1198.601587   475 0.1403025810
#> 401     401 1201.648755   487 0.1438470672
#> 402     402 1204.696051   539 0.1592065077
#> 403     403 1207.743475   461 0.1361673470
#> 404     404 1210.791027   502 0.1482776751
#> 405     405 1213.838707   452 0.1335089823
#> 406     406 1216.886515   439 0.1296691222
#> 407     407 1219.934451   486 0.1435516934
#> 408     408 1222.982515   496 0.1465054319
#> 409     409 1226.030707   447 0.1320321130
#> 410     410 1229.079027   460 0.1358719732
#> 411     411 1232.127475   473 0.1397118333
#> 412     412 1235.176051   469 0.1385303378
#> 413     413 1238.224755   443 0.1308506176
#> 414     414 1241.273587   438 0.1293737484
#> 415     415 1244.322547   454 0.1340997300
#> 416     416 1247.371635   429 0.1267153837
#> 417     417 1250.420851   431 0.1273061314
#> 418     418 1253.470195   440 0.1299644961
#> 419     419 1256.519667   404 0.1193310373
#> 420     420 1259.569267   423 0.1249431405
#> 421     421 1262.618995   386 0.1140143079
#> 422     422 1265.668851   354 0.1045623446
#> 423     423 1268.718835   399 0.1178541680
#> 424     424 1271.768947   390 0.1151958033
#> 425     425 1274.819187   348 0.1027901014
#> 426     426 1277.869555   348 0.1027901014
#> 427     427 1280.920051   338 0.0998363629
#> 428     428 1283.970675   346 0.1021993537
#> 429     429 1287.021427   344 0.1016086060
#> 430     430 1290.072307   302 0.0892029041
#> 431     431 1293.123315   357 0.1054484661
#> 432     432 1296.174450   331 0.0977687459
#> 433     433 1299.225714   309 0.0912705211
#> 434     434 1302.277106   288 0.0850676702
#> 435     435 1305.328626   327 0.0965872505
#> 436     436 1308.380274   277 0.0818185577
#> 437     437 1311.432050   304 0.0897936518
#> 438     438 1314.483954   308 0.0909751472
#> 439     439 1317.535986   284 0.0838861747
#> 440     440 1320.588146   275 0.0812278100
#> 441     441 1323.640434   261 0.0770925761
#> 442     442 1326.692850   245 0.0723665944
#> 443     443 1329.745394   286 0.0844769224
#> 444     444 1332.798066   267 0.0788648192
#> 445     445 1335.850866   246 0.0726619683
#> 446     446 1338.903794   255 0.0753203329
#> 447     447 1341.956850   240 0.0708897251
#> 448     448 1345.010034   245 0.0723665944
#> 449     449 1348.063346   233 0.0688221081
#> 450     450 1351.116786   279 0.0824093055
#> 451     451 1354.170354   240 0.0708897251
#> 452     452 1357.224050   234 0.0691174820
#> 453     453 1360.277874   238 0.0702989774
#> 454     454 1363.331826   260 0.0767972022
#> 455     455 1366.385906   241 0.0711850990
#> 456     456 1369.440114   236 0.0697082297
#> 457     457 1372.494450   245 0.0723665944
#> 458     458 1375.548914   231 0.0682313604
#> 459     459 1378.603506   240 0.0708897251
#> 460     460 1381.658226   275 0.0812278100
#> 461     461 1384.713074   249 0.0735480898
#> 462     462 1387.768050   245 0.0723665944
#> 463     463 1390.823154   249 0.0735480898
#> 464     464 1393.878385   242 0.0714804728
#> 465     465 1396.933745   221 0.0652776219
#> 466     466 1399.989233   257 0.0759110807
#> 467     467 1403.044849   237 0.0700036036
#> 468     468 1406.100593   270 0.0797509408
#> 469     469 1409.156465   235 0.0694128559
#> 470     470 1412.212465   271 0.0800463146
#> 471     471 1415.268593   279 0.0824093055
#> 472     472 1418.324849   299 0.0883167826
#> 473     473 1421.381233   321 0.0948150074
#> 474     474 1424.437745   331 0.0977687459
#> 475     475 1427.494385   307 0.0906797734
#> 476     476 1430.551153   333 0.0983594936
#> 477     477 1433.608049   383 0.1131281863
#> 478     478 1436.665073   430 0.1270107575
#> 479     479 1439.722225   428 0.1264200098
#> 480     480 1442.779505   491 0.1450285627
#> 481     481 1445.836913   487 0.1438470672
#> 482     482 1448.894449   553 0.1633417416
#> 483     483 1451.952113   570 0.1683630972
#> 484     484 1455.009905   621 0.1834271638
#> 485     485 1458.067825   709 0.2094200630
#> 486     486 1461.125873   742 0.2191674002
#> 487     487 1464.184049   794 0.2345268406
#> 488     488 1467.242353   867 0.2560891320
#> 489     489 1470.300785   851 0.2513631503
#> 490     490 1473.359345   936 0.2764699280
#> 491     491 1476.418033   947 0.2797190404
#> 492     492 1479.476849  1053 0.3110286690
#> 493     493 1482.535793  1071 0.3163453984
#> 494     494 1485.594865  1033 0.3051211919
#> 495     495 1488.654065  1044 0.3083703043
#> 496     496 1491.713392  1076 0.3178222676
#> 497     497 1494.772848   972 0.2871033868
#> 498     498 1497.832432   973 0.2873987606
#> 499     499 1500.892144  1012 0.2989183409
#> 500     500 1503.951984   918 0.2711531986
#> 501     501 1507.011952   964 0.2847403959
#> 502     502 1510.072048   824 0.2433880563
#> 503     503 1513.132272   779 0.2300962328
#> 504     504 1516.192624   734 0.2168044093
#> 505     505 1519.253104   642 0.1896300147
#> 506     506 1522.313712   604 0.1784058082
#> 507     507 1525.374448   543 0.1603880031
#> 508     508 1528.435312   476 0.1405979548
#> 509     509 1531.496304   425 0.1255338882
#> 510     510 1534.557424   414 0.1222847758
#> 511     511 1537.618672   373 0.1101744478
#> 512     512 1540.680048   326 0.0962918766
#> 513     513 1543.741552   266 0.0785694453
#> 514     514 1546.803184   249 0.0735480898
#> 515     515 1549.864944   222 0.0655729957
#> 516     516 1552.926832   220 0.0649822480
#> 517     517 1555.988848   167 0.0493274337
#> 518     518 1559.050992   200 0.0590747709
#> 519     519 1562.113264   170 0.0502135553
#> 520     520 1565.175664   178 0.0525765461
#> 521     521 1568.238192   157 0.0463736952
#> 522     522 1571.300848   150 0.0443060782
#> 523     523 1574.363632   154 0.0454875736
#> 524     524 1577.426544   124 0.0366263580
#> 525     525 1580.489584   129 0.0381032273
#> 526     526 1583.552752   130 0.0383986011
#> 527     527 1586.616048   116 0.0342633671
#> 528     528 1589.679471   146 0.0431245828
#> 529     529 1592.743023   156 0.0460783213
#> 530     530 1595.806703   134 0.0395800965
#> 531     531 1598.870511   135 0.0398754704
#> 532     532 1601.934447   126 0.0372171057
#> 533     533 1604.998511   143 0.0422384612
#> 534     534 1608.062703   123 0.0363309841
#> 535     535 1611.127023   155 0.0457829475
#> 536     536 1614.191471   141 0.0416477135
#> 537     537 1617.256047   151 0.0446014521
#> 538     538 1620.320751   122 0.0360356103
#> 539     539 1623.385583   142 0.0419430874
#> 540     540 1626.450543   132 0.0389893488
#> 541     541 1629.515631   130 0.0383986011
#> 542     542 1632.580847   113 0.0333772456
#> 543     543 1635.646191   124 0.0366263580
#> 544     544 1638.711663   158 0.0466690690
#> 545     545 1641.777263   122 0.0360356103
#> 546     546 1644.842991   145 0.0428292089
#> 547     547 1647.908847   133 0.0392847227
#> 548     548 1650.974831   118 0.0348541149
#> 549     549 1654.040943   105 0.0310142547
#> 550     550 1657.107183   106 0.0313096286
#> 551     551 1660.173551   106 0.0313096286
#> 552     552 1663.240047   121 0.0357402364
#> 553     553 1666.306671    90 0.0265836469
#> 554     554 1669.373423    99 0.0292420116
#> 555     555 1672.440303   116 0.0342633671
#> 556     556 1675.507311   100 0.0295373855
#> 557     557 1678.574447   117 0.0345587410
#> 558     558 1681.641711    90 0.0265836469
#> 559     559 1684.709103    93 0.0274697685
#> 560     560 1687.776622    96 0.0283558901
#> 561     561 1690.844270    94 0.0277651423
#> 562     562 1693.912046    97 0.0286512639
#> 563     563 1696.979950    89 0.0262882731
#> 564     564 1700.047982    79 0.0233345345
#> 565     565 1703.116142    77 0.0227437868
#> 566     566 1706.184430    71 0.0209715437
#> 567     567 1709.252846    93 0.0274697685
#> 568     568 1712.321390    78 0.0230391607
#> 569     569 1715.390062    91 0.0268790208
#> 570     570 1718.458862    93 0.0274697685
#> 571     571 1721.527790    86 0.0254021515
#> 572     572 1724.596846   109 0.0321957502
#> 573     573 1727.666030    87 0.0256975254
#> 574     574 1730.735342    97 0.0286512639
#> 575     575 1733.804782    96 0.0283558901
#> 576     576 1736.874350    94 0.0277651423
#> 577     577 1739.944046   109 0.0321957502
#> 578     578 1743.013870    97 0.0286512639
#> 579     579 1746.083822   113 0.0333772456
#> 580     580 1749.153902   116 0.0342633671
#> 581     581 1752.224110   103 0.0304235070
#> 582     582 1755.294446   112 0.0330818717
#> 583     583 1758.364910   124 0.0366263580
#> 584     584 1761.435502   130 0.0383986011
#> 585     585 1764.506222   127 0.0375124795
#> 586     586 1767.577070   108 0.0319003763
#> 587     587 1770.648046   133 0.0392847227
#> 588     588 1773.719150   120 0.0354448626
#> 589     589 1776.790382   130 0.0383986011
#> 590     590 1779.861742   138 0.0407615919
#> 591     591 1782.933230   130 0.0383986011
#> 592     592 1786.004845   157 0.0463736952
#> 593     593 1789.076589   146 0.0431245828
#> 594     594 1792.148461   129 0.0381032273
#> 595     595 1795.220461   142 0.0419430874
#> 596     596 1798.292589   153 0.0451921998
#> 597     597 1801.364845   120 0.0354448626
#> 598     598 1804.437229   149 0.0440107043
#> 599     599 1807.509741   147 0.0434199566
#> 600     600 1810.582381   134 0.0395800965
#> 601     601 1813.655149   113 0.0333772456
#> 602     602 1816.728045   122 0.0360356103
#> 603     603 1819.801069   114 0.0336726194
#> 604     604 1822.874221   114 0.0336726194
#> 605     605 1825.947501   128 0.0378078534
#> 606     606 1829.020909   120 0.0354448626
#> 607     607 1832.094445    99 0.0292420116
#> 608     608 1835.168109   101 0.0298327593
#> 609     609 1838.241901   119 0.0351494887
#> 610     610 1841.315821    92 0.0271743946
#> 611     611 1844.389869   112 0.0330818717
#> 612     612 1847.464045    88 0.0259928992
#> 613     613 1850.538349    84 0.0248114038
#> 614     614 1853.612781    97 0.0286512639
#> 615     615 1856.687341    66 0.0194946744
#> 616     616 1859.762029    98 0.0289466378
#> 617     617 1862.836845    69 0.0203807960
#> 618     618 1865.911789    75 0.0221530391
#> 619     619 1868.986861    66 0.0194946744
#> 620     620 1872.062061    87 0.0256975254
#> 621     621 1875.137389    63 0.0186085528
#> 622     622 1878.212845    79 0.0233345345
#> 623     623 1881.288429    55 0.0162455620
#> 624     624 1884.364140    79 0.0233345345
#> 625     625 1887.439980    62 0.0183131790
#> 626     626 1890.515948    65 0.0191993006
#> 627     627 1893.592044    71 0.0209715437
#> 628     628 1896.668268    59 0.0174270574
#> 629     629 1899.744620    72 0.0212669175
#> 630     630 1902.821100    71 0.0209715437
#> 631     631 1905.897708    56 0.0165409359
#> 632     632 1908.974444    70 0.0206761698
#> 633     633 1912.051308    75 0.0221530391
#> 634     634 1915.128300    63 0.0186085528
#> 635     635 1918.205420    68 0.0200854221
#> 636     636 1921.282668    54 0.0159501882
#> 637     637 1924.360044    72 0.0212669175
#> 638     638 1927.437548    51 0.0150640666
#> 639     639 1930.515180    52 0.0153594404
#> 640     640 1933.592940    73 0.0215622914
#> 641     641 1936.670828    50 0.0147686927
#> 642     642 1939.748844    48 0.0141779450
#> 643     643 1942.826988    62 0.0183131790
#> 644     644 1945.905260    56 0.0165409359
#> 645     645 1948.983660    58 0.0171316836
#> 646     646 1952.062188    60 0.0177224313
#> 647     647 1955.140844    45 0.0132918235
#> 648     648 1958.219628    76 0.0224484130
#> 649     649 1961.298540    42 0.0124057019
#> 650     650 1964.377580    62 0.0183131790
#> 651     651 1967.456748    62 0.0183131790
#> 652     652 1970.536044    55 0.0162455620
#> 653     653 1973.615468    49 0.0144733189
#> 654     654 1976.695020    48 0.0141779450
#> 655     655 1979.774700    55 0.0162455620
#> 656     656 1982.854507    43 0.0127010758
#> 657     657 1985.934443    58 0.0171316836
#> 658     658 1989.014507    55 0.0162455620
#> 659     659 1992.094699    51 0.0150640666
#> 660     660 1995.175019    72 0.0212669175
#> 661     661 1998.255467    55 0.0162455620
#> 662     662 2001.336043    68 0.0200854221
#> 663     663 2004.416747    60 0.0177224313
#> 664     664 2007.497579    51 0.0150640666
#> 665     665 2010.578539    35 0.0103380849
#> 666     666 2013.659627    55 0.0162455620
#> 667     667 2016.740843    58 0.0171316836
#> 668     668 2019.822187    53 0.0156548143
#> 669     669 2022.903659    53 0.0156548143
#> 670     670 2025.985259    50 0.0147686927
#> 671     671 2029.066987    39 0.0115195803
#> 672     672 2032.148843    45 0.0132918235
#> 673     673 2035.230827    42 0.0124057019
#> 674     674 2038.312939    55 0.0162455620
#> 675     675 2041.395179    47 0.0138825712
#> 676     676 2044.477547    60 0.0177224313
#> 677     677 2047.560043    48 0.0141779450
#> 678     678 2050.642667    56 0.0165409359
#> 679     679 2053.725419    51 0.0150640666
#> 680     680 2056.808299    62 0.0183131790
#> 681     681 2059.891307    55 0.0162455620
#> 682     682 2062.974443    60 0.0177224313
#> 683     683 2066.057707    48 0.0141779450
#> 684     684 2069.141099    48 0.0141779450
#> 685     685 2072.224619    68 0.0200854221
#> 686     686 2075.308267    49 0.0144733189
#> 687     687 2078.392042    51 0.0150640666
#> 688     688 2081.475946    44 0.0129964496
#> 689     689 2084.559978    48 0.0141779450
#> 690     690 2087.644138    52 0.0153594404
#> 691     691 2090.728426    52 0.0153594404
#> 692     692 2093.812842    61 0.0180178051
#> 693     693 2096.897386    55 0.0162455620
#> 694     694 2099.982058    57 0.0168363097
#> 695     695 2103.066858    65 0.0191993006
#> 696     696 2106.151786    69 0.0203807960
#> 697     697 2109.236842    47 0.0138825712
#> 698     698 2112.322026    57 0.0168363097
#> 699     699 2115.407338    63 0.0186085528
#> 700     700 2118.492778    60 0.0177224313
#> 701     701 2121.578346    63 0.0186085528
#> 702     702 2124.664042    73 0.0215622914
#> 703     703 2127.749866    70 0.0206761698
#> 704     704 2130.835818    77 0.0227437868
#> 705     705 2133.921898    74 0.0218576652
#> 706     706 2137.008106    68 0.0200854221
#> 707     707 2140.094442    80 0.0236299084
#> 708     708 2143.180906    78 0.0230391607
#> 709     709 2146.267498    89 0.0262882731
#> 710     710 2149.354218    71 0.0209715437
#> 711     711 2152.441066    73 0.0215622914
#> 712     712 2155.528042    56 0.0165409359
#> 713     713 2158.615146    90 0.0265836469
#> 714     714 2161.702378    81 0.0239252822
#> 715     715 2164.789738    78 0.0230391607
#> 716     716 2167.877226    89 0.0262882731
#> 717     717 2170.964842    81 0.0239252822
#> 718     718 2174.052586    95 0.0280605162
#> 719     719 2177.140457   109 0.0321957502
#> 720     720 2180.228457    74 0.0218576652
#> 721     721 2183.316585    86 0.0254021515
#> 722     722 2186.404841    95 0.0280605162
#> 723     723 2189.493225    78 0.0230391607
#> 724     724 2192.581737    65 0.0191993006
#> 725     725 2195.670377    80 0.0236299084
#> 726     726 2198.759145    78 0.0230391607
#> 727     727 2201.848041    84 0.0248114038
#> 728     728 2204.937065    83 0.0245160299
#> 729     729 2208.026217    68 0.0200854221
#> 730     730 2211.115497    82 0.0242206561
#> 731     731 2214.204905    76 0.0224484130
#> 732     732 2217.294441    67 0.0197900483
#> 733     733 2220.384105    69 0.0203807960
#> 734     734 2223.473897    66 0.0194946744
#> 735     735 2226.563817    68 0.0200854221
#> 736     736 2229.653865    74 0.0218576652
#> 737     737 2232.744041    71 0.0209715437
#> 738     738 2235.834345    72 0.0212669175
#> 739     739 2238.924777    62 0.0183131790
#> 740     740 2242.015337    69 0.0203807960
#> 741     741 2245.106025    80 0.0236299084
#> 742     742 2248.196841    71 0.0209715437
#> 743     743 2251.287785    49 0.0144733189
#> 744     744 2254.378857    52 0.0153594404
#> 745     745 2257.470057    60 0.0177224313
#> 746     746 2260.561385    68 0.0200854221
#> 747     747 2263.652841    69 0.0203807960
#> 748     748 2266.744425    54 0.0159501882
#> 749     749 2269.836137    57 0.0168363097
#> 750     750 2272.927977    60 0.0177224313
#> 751     751 2276.019944    62 0.0183131790
#> 752     752 2279.112040    65 0.0191993006
#> 753     753 2282.204264    68 0.0200854221
#> 754     754 2285.296616    73 0.0215622914
#> 755     755 2288.389096    70 0.0206761698
#> 756     756 2291.481704    52 0.0153594404
#> 757     757 2294.574440    63 0.0186085528
#> 758     758 2297.667304    62 0.0183131790
#> 759     759 2300.760296    52 0.0153594404
#> 760     760 2303.853416    54 0.0159501882
#> 761     761 2306.946664    43 0.0127010758
#> 762     762 2310.040040    54 0.0159501882
#> 763     763 2313.133544    58 0.0171316836
#> 764     764 2316.227176    66 0.0194946744
#> 765     765 2319.320936    61 0.0180178051
#> 766     766 2322.414824    59 0.0174270574
#> 767     767 2325.508840    55 0.0162455620
#> 768     768 2328.602984    67 0.0197900483
#> 769     769 2331.697256    73 0.0215622914
#> 770     770 2334.791656    65 0.0191993006
#> 771     771 2337.886184    59 0.0174270574
#> 772     772 2340.980840    59 0.0174270574
#> 773     773 2344.075624    63 0.0186085528
#> 774     774 2347.170536    48 0.0141779450
#> 775     775 2350.265576    34 0.0100427111
#> 776     776 2353.360744    50 0.0147686927
#> 777     777 2356.456040    46 0.0135871973
#> 778     778 2359.551464    48 0.0141779450
#> 779     779 2362.647016    61 0.0180178051
#> 780     780 2365.742696    47 0.0138825712
#> 781     781 2368.838504    53 0.0156548143
#> 782     782 2371.934440    40 0.0118149542
#> 783     783 2375.030503    38 0.0112242065
#> 784     784 2378.126695    36 0.0106334588
#> 785     785 2381.223015    51 0.0150640666
#> 786     786 2384.319463    45 0.0132918235
#> 787     787 2387.416039    45 0.0132918235
#> 788     788 2390.512743    37 0.0109288326
#> 789     789 2393.609575    47 0.0138825712
#> 790     790 2396.706535    39 0.0115195803
#> 791     791 2399.803623    45 0.0132918235
#> 792     792 2402.900839    38 0.0112242065
#> 793     793 2405.998183    53 0.0156548143
#> 794     794 2409.095655    39 0.0115195803
#> 795     795 2412.193255    45 0.0132918235
#> 796     796 2415.290983    42 0.0124057019
#> 797     797 2418.388839    41 0.0121103280
#> 798     798 2421.486823    35 0.0103380849
#> 799     799 2424.584935    31 0.0091565895
#> 800     800 2427.683175    38 0.0112242065
#> 801     801 2430.781543    31 0.0091565895
#> 802     802 2433.880039    34 0.0100427111
#> 803     803 2436.978663    43 0.0127010758
#> 804     804 2440.077415    25 0.0073843464
#> 805     805 2443.176295    27 0.0079750941
#> 806     806 2446.275303    28 0.0082704679
#> 807     807 2449.374439    28 0.0082704679
#> 808     808 2452.473703    29 0.0085658418
#> 809     809 2455.573095    32 0.0094519634
#> 810     810 2458.672615    52 0.0153594404
#> 811     811 2461.772263    36 0.0106334588
#> 812     812 2464.872039    34 0.0100427111
#> 813     813 2467.971943    31 0.0091565895
#> 814     814 2471.071975    25 0.0073843464
#> 815     815 2474.172134    27 0.0079750941
#> 816     816 2477.272422    27 0.0079750941
#> 817     817 2480.372838    34 0.0100427111
#> 818     818 2483.473382    22 0.0064982248
#> 819     819 2486.574054    26 0.0076797202
#> 820     820 2489.674854    24 0.0070889725
#> 821     821 2492.775782    28 0.0082704679
#> 822     822 2495.876838    26 0.0076797202
#> 823     823 2498.978022    19 0.0056121032
#> 824     824 2502.079334    14 0.0041352340
#> 825     825 2505.180774    23 0.0067935987
#> 826     826 2508.282342    18 0.0053167294
#> 827     827 2511.384038    31 0.0091565895
#> 828     828 2514.485862    19 0.0056121032
#> 829     829 2517.587814    25 0.0073843464
#> 830     830 2520.689894    19 0.0056121032
#> 831     831 2523.792102    26 0.0076797202
#> 832     832 2526.894438    14 0.0041352340
#> 833     833 2529.996902    18 0.0053167294
#> 834     834 2533.099494    18 0.0053167294
#> 835     835 2536.202214    21 0.0062028509
#> 836     836 2539.305062    25 0.0073843464
#> 837     837 2542.408038    15 0.0044306078
#> 838     838 2545.511142    25 0.0073843464
#> 839     839 2548.614374    10 0.0029537385
#> 840     840 2551.717734    21 0.0062028509
#> 841     841 2554.821222    16 0.0047259817
#> 842     842 2557.924838    19 0.0056121032
#> 843     843 2561.028582    26 0.0076797202
#> 844     844 2564.132454    19 0.0056121032
#> 845     845 2567.236454    19 0.0056121032
#> 846     846 2570.340581    26 0.0076797202
#> 847     847 2573.444837    18 0.0053167294
#> 848     848 2576.549221    17 0.0050213555
#> 849     849 2579.653733    20 0.0059074771
#> 850     850 2582.758373    22 0.0064982248
#> 851     851 2585.863141    30 0.0088612156
#> 852     852 2588.968037    38 0.0112242065
#> 853     853 2592.073061    24 0.0070889725
#> 854     854 2595.178213    32 0.0094519634
#> 855     855 2598.283493    36 0.0106334588
#> 856     856 2601.388901    42 0.0124057019
#> 857     857 2604.494437    43 0.0127010758
#> 858     858 2607.600101    34 0.0100427111
#> 859     859 2610.705893    64 0.0189039267
#> 860     860 2613.811813    48 0.0141779450
#> 861     861 2616.917861    59 0.0174270574
#> 862     862 2620.024037    73 0.0215622914
#> 863     863 2623.130341    67 0.0197900483
#> 864     864 2626.236773    80 0.0236299084
#> 865     865 2629.343333    78 0.0230391607
#> 866     866 2632.450021    83 0.0245160299
#> 867     867 2635.556837    98 0.0289466378
#> 868     868 2638.663781    79 0.0233345345
#> 869     869 2641.770853    84 0.0248114038
#> 870     870 2644.878053   102 0.0301281332
#> 871     871 2647.985381    91 0.0268790208
#> 872     872 2651.092837   100 0.0295373855
#> 873     873 2654.200421    94 0.0277651423
#> 874     874 2657.308133    91 0.0268790208
#> 875     875 2660.415973   114 0.0336726194
#> 876     876 2663.523941   116 0.0342633671
#> 877     877 2666.632037   101 0.0298327593
#> 878     878 2669.740260    93 0.0274697685
#> 879     879 2672.848612   105 0.0310142547
#> 880     880 2675.957092   114 0.0336726194
#> 881     881 2679.065700   102 0.0301281332
#> 882     882 2682.174436   111 0.0327864979
#> 883     883 2685.283300    96 0.0283558901
#> 884     884 2688.392292    96 0.0283558901
#> 885     885 2691.501412    93 0.0274697685
#> 886     886 2694.610660    90 0.0265836469
#> 887     887 2697.720036    75 0.0221530391
#> 888     888 2700.829540    69 0.0203807960
#> 889     889 2703.939172    51 0.0150640666
#> 890     890 2707.048932    73 0.0215622914
#> 891     891 2710.158820    73 0.0215622914
#> 892     892 2713.268836    63 0.0186085528
#> 893     893 2716.378980    57 0.0168363097
#> 894     894 2719.489252    55 0.0162455620
#> 895     895 2722.599652    48 0.0141779450
#> 896     896 2725.710180    40 0.0118149542
#> 897     897 2728.820836    39 0.0115195803
#> 898     898 2731.931620    44 0.0129964496
#> 899     899 2735.042532    44 0.0129964496
#> 900     900 2738.153572    35 0.0103380849
#> 901     901 2741.264740    25 0.0073843464
#> 902     902 2744.376036    35 0.0103380849
#> 903     903 2747.487460    26 0.0076797202
#> 904     904 2750.599012    16 0.0047259817
#> 905     905 2753.710692    17 0.0050213555
#> 906     906 2756.822500    15 0.0044306078
#> 907     907 2759.934436    15 0.0044306078
#> 908     908 2763.046500    11 0.0032491124
#> 909     909 2766.158692    15 0.0044306078
#> 910     910 2769.271011    13 0.0038398601
#> 911     911 2772.383459    10 0.0029537385
#> 912     912 2775.496035    13 0.0038398601
#> 913     913 2778.608739     9 0.0026583647
#> 914     914 2781.721571     7 0.0020676170
#> 915     915 2784.834531     4 0.0011814954
#> 916     916 2787.947619    12 0.0035444863
#> 917     917 2791.060835     5 0.0014768693
#> 918     918 2794.174179     6 0.0017722431
#> 919     919 2797.287651     5 0.0014768693
#> 920     920 2800.401251     4 0.0011814954
#> 921     921 2803.514979     2 0.0005907477
#> 922     922 2806.628835     3 0.0008861216
#> 923     923 2809.742819     2 0.0005907477
#> 924     924 2812.856931     4 0.0011814954
#> 925     925 2815.971171     2 0.0005907477
#> 926     926 2819.085539     4 0.0011814954
#> 927     927 2822.200035     2 0.0005907477
#> 928     928 2825.314659     7 0.0020676170
#> 929     929 2828.429411     5 0.0014768693
#> 930     930 2831.544291     2 0.0005907477
#> 931     931 2834.659299     1 0.0002953739
#> 932     932 2837.774435     3 0.0008861216
#> 933     933 2840.889699     4 0.0011814954
#> 934     934 2844.005091     1 0.0002953739
#> 935     935 2847.120611     2 0.0005907477
#> 936     936 2850.236259     1 0.0002953739
#> 937     937 2853.352035     1 0.0002953739
#> 938     938 2856.467939     4 0.0011814954
#> 939     939 2859.583971     0 0.0000000000
#> 940     940 2862.700131     5 0.0014768693
#> 941     941 2865.816419     2 0.0005907477
#> 942     942 2868.932834     2 0.0005907477
#> 943     943 2872.049378     2 0.0005907477
#> 944     944 2875.166050     0 0.0000000000
#> 945     945 2878.282850     0 0.0000000000
#> 946     946 2881.399778     1 0.0002953739
#> 947     947 2884.516834     3 0.0008861216
#> 948     948 2887.634018     4 0.0011814954
#> 949     949 2890.751330     1 0.0002953739
#> 950     950 2893.868770     3 0.0008861216
#> 951     951 2896.986338     2 0.0005907477
#> 952     952 2900.104034     4 0.0011814954
#> 953     953 2903.221858     2 0.0005907477
#> 954     954 2906.339810     2 0.0005907477
#> 955     955 2909.457890     0 0.0000000000
#> 956     956 2912.576098     5 0.0014768693
#> 957     957 2915.694434     2 0.0005907477
#> 958     958 2918.812898     2 0.0005907477
#> 959     959 2921.931490     2 0.0005907477
#> 960     960 2925.050210     1 0.0002953739
#> 961     961 2928.169058     4 0.0011814954
#> 962     962 2931.288034     2 0.0005907477
#> 963     963 2934.407138     3 0.0008861216
#> 964     964 2937.526370     2 0.0005907477
#> 965     965 2940.645730     2 0.0005907477
#> 966     966 2943.765218     2 0.0005907477
#> 967     967 2946.884834     3 0.0008861216
#> 968     968 2950.004578     2 0.0005907477
#> 969     969 2953.124450     2 0.0005907477
#> 970     970 2956.244450     1 0.0002953739
#> 971     971 2959.364578     2 0.0005907477
#> 972     972 2962.484834     2 0.0005907477
#> 973     973 2965.605218     1 0.0002953739
#> 974     974 2968.725729     1 0.0002953739
#> 975     975 2971.846369     2 0.0005907477
#> 976     976 2974.967137     2 0.0005907477
#> 977     977 2978.088033     2 0.0005907477
#> 978     978 2981.209057     2 0.0005907477
#> 979     979 2984.330209     1 0.0002953739
#> 980     980 2987.451489     0 0.0000000000
#> 981     981 2990.572897     3 0.0008861216
#> 982     982 2993.694433     0 0.0000000000
#> 983     983 2996.816097     2 0.0005907477
#> 984     984 2999.937889     0 0.0000000000
#> 985     985 3003.059809     1 0.0002953739
#> 986     986 3006.181857     2 0.0005907477
#> 987     987 3009.304033     0 0.0000000000
#> 988     988 3012.426337     1 0.0002953739
#> 989     989 3015.548769     2 0.0005907477
#> 990     990 3018.671329     3 0.0008861216
#> 991     991 3021.794017     1 0.0002953739
#> 992     992 3024.916833     1 0.0002953739
#> 993     993 3028.039777     0 0.0000000000
#> 994     994 3031.162849     0 0.0000000000
#> 995     995 3034.286049     1 0.0002953739
#> 996     996 3037.409377     1 0.0002953739
#> 997     997 3040.532833     1 0.0002953739
#> 998     998 3043.656417     0 0.0000000000
#> 999     999 3046.780129     0 0.0000000000
#> 1000   1000 3049.903969     1 0.0002953739
#> 1001   1001 3053.027937     0 0.0000000000
#> 1002   1002 3056.152033     0 0.0000000000
#> 1003   1003 3059.276257     2 0.0005907477
#> 1004   1004 3062.400609     0 0.0000000000
#> 1005   1005 3065.525088     0 0.0000000000
#> 1006   1006 3068.649696     3 0.0008861216
#> 1007   1007 3071.774432     3 0.0008861216
#> 1008   1008 3074.899296     3 0.0008861216
#> 1009   1009 3078.024288     1 0.0002953739
#> 1010   1010 3081.149408     1 0.0002953739
#> 1011   1011 3084.274656     2 0.0005907477
#> 1012   1012 3087.400032     0 0.0000000000
#> 1013   1013 3090.525536     1 0.0002953739
#> 1014   1014 3093.651168     1 0.0002953739
#> 1015   1015 3096.776928     1 0.0002953739
#> 1016   1016 3099.902816     1 0.0002953739
#> 1017   1017 3103.028832     0 0.0000000000
#> 1018   1018 3106.154976     1 0.0002953739
#> 1019   1019 3109.281248     1 0.0002953739
#> 1020   1020 3112.407648     1 0.0002953739
#> 1021   1021 3115.534176     2 0.0005907477
#> 1022   1022 3118.660832     0 0.0000000000
#> 1023   1023 3121.787616     1 0.0002953739
#> 1024   1024 3124.914528     0 0.0000000000

# Subset
spectrum[["hash"]]

#> [1] "e0a2c67173cf7f407db7148ae0058bbf"

spectrum[["name"]]

#> [1] "test_CNF"

spectrum[["date"]]

#> [1] "2019-02-07 11:48:18 UTC"

spectrum[["instrument"]]

#> [1] "InSpector 1000"

spectrum[["file_format"]]

#> [1] "CNF"

spectrum[["live_time"]]

#> [1] 3385.54

spectrum[["real_time"]]

#> [1] 3403.67

spectrum[["calibration"]]

#> 
#> Call:
#> stats::lm(formula = 0 ~ 0)
#> 
#> No coefficients
#>