LLS operator

LLS operator

LLS(x)

inverseLLS(x)

Arguments

x
A numeric vector.

Value

A numeric vector.

References

Morháč, M., Kliman, J., Matoušek, V., Veselský, M. and Turzo, I. (1997). Background elimination methods for multidimensional gamma-ray spectra. Nuclear Instruments and Methods in Physics Research Section A: Accelerators, Spectrometers, Detectors and Associated Equipment, 401(1), p. 113-132. DOI: 10.1016/S0168-9002(97)01023-1

Morháč, M. and Matoušek, V. (2008). Peak Clipping Algorithms for Background Estimation in Spectroscopic Data. Applied Spectroscopy, 62(1), p. 91-106. DOI: 10.1366/000370208783412762