Get or Set Parts of an Object

Getters and setters to extract or replace parts of an object.

get_hash(x)

get_names(x)

set_names(x) <- value

get_livetime(x)

get_realtime(x)

get_channels(x)

get_counts(x)

get_rates(x)

get_energy(x, ...)

set_energy(x, ...) <- value

get_method(x)

set_method(x) <- value

get_residuals(x)

range_channels(x, ...)

range_energy(x, ...)

# S4 method for GammaSpectrum
length(x)

# S4 method for GammaSpectrum
get_hash(x)

# S4 method for GammaSpectrum
get_names(x)

# S4 method for GammaSpectrum
get_livetime(x)

# S4 method for GammaSpectrum
get_realtime(x)

# S4 method for GammaSpectrum
get_channels(x)

# S4 method for GammaSpectrum
get_counts(x)

# S4 method for GammaSpectrum
get_rates(x)

# S4 method for GammaSpectrum
get_energy(x)

# S4 method for GammaSpectrum
range_energy(x, na.rm = FALSE)

# S4 method for GammaSpectrum
range_channels(x, na.rm = FALSE)

# S4 method for Baseline
get_method(x)

# S4 method for GammaSpectra
get_hash(x)

# S4 method for GammaSpectra
get_names(x)

# S4 method for GammaSpectra
get_livetime(x)

# S4 method for GammaSpectra
get_realtime(x)

# S4 method for GammaSpectra
get_channels(x)

# S4 method for GammaSpectra
get_counts(x)

# S4 method for GammaSpectra
get_rates(x)

# S4 method for GammaSpectra
get_energy(x)

# S4 method for GammaSpectra
range_energy(x, na.rm = FALSE)

# S4 method for GammaSpectra
range_channels(x, na.rm = FALSE)

# S4 method for DoseRateModel
get_residuals(x)

# S4 method for PeakPosition
get_hash(x)

# S4 method for PeakPosition
get_channels(x)

# S4 method for PeakPosition
get_energy(x, expected = FALSE)

# S4 method for GammaSpectrum
set_names(x) <- value

# S4 method for Baseline
set_method(x) <- value

# S4 method for GammaSpectra
set_names(x) <- value

# S4 method for PeakPosition,numeric
set_energy(x, expected = TRUE) <- value

Arguments

x
An object from which to get or set element(s).

value
A possible value for the element(s) of x.

...
Currently not used.

na.rm
A logical scalar: should NA be omitted?

expected
TODO.

Value

An object of the same sort as x with the new values assigned.

See also

Author

N. Frerebeau