Common Operations on GammaSpectrum Objects

Performs common operations on GammaSpectrum objects.

# S4 method for GammaSpectrum,GammaSpectrum
Arith(e1, e2)

# S4 method for GammaSpectrum,numeric
Arith(e1, e2)

# S4 method for GammaSpectrum,GammaSpectrum
Compare(e1, e2)

# S4 method for GammaSpectrum,numeric
Compare(e1, e2)

# S4 method for GammaSpectrum,GammaSpectrum
Logic(e1, e2)

# S4 method for GammaSpectrum,numeric
Logic(e1, e2)

# S4 method for GammaSpectrum,logical
Logic(e1, e2)

# S4 method for GammaSpectrum
Math(x)

# S4 method for GammaSpectrum
Math2(x, digits)

# S4 method for GammaSpectrum
Summary(x, ..., na.rm = FALSE)

Arguments

x, e1, e2

An object (typically a GammaSpectrum object).

digits

A length-one numeric vector giving the number of digits to be used in round or signif.

...

Further arguments passed to or from methods.

na.rm

A logical scalar: should missing values (including NaN) be omitted from the calculations?

Group Generics

GammaSpectrum objects have support for S4 group generic functionality to operate within elements across objects:

Arith

"+", "-", "*", "^", "%%", "%/%", "/"

Compare

"==", ">", "<", "!=", "<=", ">="

Logic

"&", "|"

Math

"abs", "sign", "sqrt", "ceiling", "floor", "trunc", "cummax", "cummin", "cumprod", "cumsum", "log", "log10", "log2", "log1p", "acos", "acosh", "asin", "asinh", "atan", "atanh", "exp", "expm1", "cos", "cosh", "cospi", "sin", "sinh", "sinpi", "tan", "tanh", "tanpi", "gamma", "lgamma", "digamma", "trigamma"

Math2

"round", "signif"

Ops

"Arith", "Compare", "Logic"

Summary

"min", "max", "range", "prod", "sum", "any", "all"

Author

N. Frerebeau

Examples

## No examples