Transform Intensities

Transform Intensities

Usage

signal_stabilize(object, ...)

# S4 method for class 'GammaSpectrum'
signal_stabilize(object, f, ...)

# S4 method for class 'GammaSpectra'
signal_stabilize(object, f, ...)

Arguments

object

A GammaSpectrum object.

...

Extra arguments to be passed to f.

f

A function that takes a numeric vector as argument and returns a numeric vector.

Value

A GammaSpectrum or GammaSpectra object with transformed intensities.

Details

The stabilization step aims at improving the identification of peaks with a low signal-to-noise ratio. This particularly targets higher energy peaks.

See also

Author

N. Frerebeau