An S4 class to represent a collection of gamma sectra

Represents a collection of spectra of gamma ray spectrometry measurements.

# S4 method for GammaSpectra
[(x, i, j)

# S4 method for GammaSpectra
[[(x, i)

Arguments

x
An object of class GammaSpectrum.

i, j
Indices specifying elements to extract or replace (see below). Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object.

Details

This class extends the base list and can only contains GammaSpectrum objects.

Methods

estimateBaseline

Estimate the baseline of each gamma spectrum in a GammaSpectra object. See estimateBaseline for details.

estimateDoseRate

Estimate the in-situ gamma dose rate of each gamma spectrum in a GammaSpectra object. See estimateDoseRate for details.

findPeaks

Look for local maxima to extract peaks out of each gamma spectrum in a GammaSpectra object. See findPeaks for details.

removeBaseline

Estimate and remove the baseline of each gamma spectrum in a GammaSpectra object. See removeBaseline for details.

Access

In the code snippets below, x is a GammaSpectra object.

length(x)

Get the number of elements in x.

names(x), names(x) <- value

Get or set the names of the elements according to value.

Coerce

In the code snippets below, x is a GammaSpectra object.

as(x, "list")

Coerces x to a list.

Subset

In the code snippets below, x is a GammaSpectra object.

x[i]

Extracts the elements selected by subscript i. i can be missing or NULL, numeric or character vector or a factor. Returns a new GammaSpectra object.

x[i, j]

Like the above but allows to select a slot thru j (see examples). j is a character vector of length one. Returns a list.

x[[i]]

Extracts the elements selected by subscript i. i can be a numeric or character vector of length one. Returns the corresponding GammaSpectrum object.

See also

Examples

# Import all CNF files in a given directory
spc_dir <- system.file("extdata/cerege/", package = "gamma")
(spectra <- read(spc_dir))

#> A collection of 6 gamma spectra: BRIQUE, C341, C347, GOU, PB, PEP

# Access
names(spectra)

#> [1] "BRIQUE" "C341"   "C347"   "GOU"    "PB"     "PEP"   

length(spectra)

#> [1] 6

# Coerce
as(spectra, "list")

#> $BRIQUE
#> Gamma spectrum:
#>   Reference: BRIQUE
#>   Instrument: InSpector 1000
#>   Date: 2019-03-27
#>   Number of chanels: 1024
#>   Energy range (keV): -7-3124.91
#> 
#> $C341
#> Gamma spectrum:
#>   Reference: C341
#>   Instrument: InSpector 1000
#>   Date: 2019-02-06
#>   Number of chanels: 1024
#>   Energy range (keV): -7-3124.91
#> 
#> $C347
#> Gamma spectrum:
#>   Reference: C347
#>   Instrument: InSpector 1000
#>   Date: 2019-02-06
#>   Number of chanels: 1024
#>   Energy range (keV): -7-3124.91
#> 
#> $GOU
#> Gamma spectrum:
#>   Reference: GOU
#>   Instrument: InSpector 1000
#>   Date: 2019-02-07
#>   Number of chanels: 1024
#>   Energy range (keV): -7-3124.91
#> 
#> $PB
#> Gamma spectrum:
#>   Reference: PB
#>   Instrument: InSpector 1000
#>   Date: 2019-03-27
#>   Number of chanels: 1024
#>   Energy range (keV): -7-3124.91
#> 
#> $PEP
#> Gamma spectrum:
#>   Reference: PEP
#>   Instrument: InSpector 1000
#>   Date: 2019-02-07
#>   Number of chanels: 1024
#>   Energy range (keV): -7-3124.91
#> 

# Subset
spectra[] # All spectra

#> A collection of 6 gamma spectra: BRIQUE, C341, C347, GOU, PB, PEP

spectra[NULL] # All spectra

#> A collection of 6 gamma spectra: BRIQUE, C341, C347, GOU, PB, PEP

spectra[1] # The first spectrum

#> A collection of 1 gamma spectrum: BRIQUE

spectra[-6] # Delete the sixth spectrum

#> A collection of 5 gamma spectra: BRIQUE, C341, C347, GOU, PB

spectra[1:3] # The first three spectra

#> A collection of 3 gamma spectra: BRIQUE, C341, C347

spectra[c(1, 3)] # The first and third spectra

#> A collection of 2 gamma spectra: BRIQUE, C347

spectra["BRIQUE"] # The spectrum named 'BRIQUE'

#> A collection of 1 gamma spectrum: BRIQUE

spectra[c("BRIQUE", "C347")] # The spectra named 'BRIQUE' and 'C347'

#> A collection of 2 gamma spectra: BRIQUE, C347

spectra[1:3, "energy"] # The slot 'energy' of the first three spectra

#> $BRIQUE
#>    [1]   -7.004032   -4.007936   -1.011712    1.984640    4.981120    7.977728
#>    [7]   10.974464   13.971328   16.968320   19.965440   22.962688   25.960064
#>   [13]   28.957568   31.955200   34.952960   37.950848   40.948863   43.947007
#>   [19]   46.945279   49.943679   52.942207   55.940863   58.939647   61.938559
#>   [25]   64.937599   67.936767   70.936063   73.935487   76.935039   79.934719
#>   [31]   82.934527   85.934463   88.934527   91.934719   94.935039   97.935487
#>   [37]  100.936063  103.936767  106.937599  109.938559  112.939647  115.940863
#>   [43]  118.942207  121.943679  124.945279  127.947007  130.948863  133.950847
#>   [49]  136.952958  139.955198  142.957566  145.960062  148.962686  151.965438
#>   [55]  154.968318  157.971326  160.974462  163.977726  166.981118  169.984638
#>   [61]  172.988286  175.992062  178.995966  181.999998  185.004158  188.008446
#>   [67]  191.012862  194.017406  197.022078  200.026878  203.031806  206.036862
#>   [73]  209.042046  212.047358  215.052798  218.058366  221.064062  224.069886
#>   [79]  227.075838  230.081918  233.088125  236.094461  239.100925  242.107517
#>   [85]  245.114237  248.121085  251.128061  254.135165  257.142397  260.149757
#>   [91]  263.157245  266.164861  269.172605  272.180477  275.188477  278.196605
#>   [97]  281.204861  284.213245  287.221757  290.230397  293.239165  296.248061
#>  [103]  299.257085  302.266237  305.275517  308.284925  311.294461  314.304125
#>  [109]  317.313917  320.323837  323.333885  326.344061  329.354364  332.364796
#>  [115]  335.375356  338.386044  341.396860  344.407804  347.418876  350.430076
#>  [121]  353.441404  356.452860  359.464444  362.476156  365.487996  368.499964
#>  [127]  371.512060  374.524284  377.536636  380.549116  383.561724  386.574460
#>  [133]  389.587324  392.600316  395.613436  398.626684  401.640060  404.653564
#>  [139]  407.667196  410.680956  413.694844  416.708860  419.723004  422.737276
#>  [145]  425.751675  428.766203  431.780859  434.795643  437.810555  440.825595
#>  [151]  443.840763  446.856059  449.871483  452.887035  455.902715  458.918523
#>  [157]  461.934459  464.950523  467.966715  470.983035  473.999483  477.016059
#>  [163]  480.032763  483.049595  486.066555  489.083643  492.100859  495.118203
#>  [169]  498.135675  501.153275  504.171003  507.188859  510.206843  513.224955
#>  [175]  516.243195  519.261563  522.280058  525.298682  528.317434  531.336314
#>  [181]  534.355322  537.374458  540.393722  543.413114  546.432634  549.452282
#>  [187]  552.472058  555.491962  558.511994  561.532154  564.552442  567.572858
#>  [193]  570.593402  573.614074  576.634874  579.655802  582.676858  585.698042
#>  [199]  588.719354  591.740794  594.762362  597.784058  600.805882  603.827834
#>  [205]  606.849914  609.872122  612.894458  615.916922  618.939513  621.962233
#>  [211]  624.985081  628.008057  631.031161  634.054393  637.077753  640.101241
#>  [217]  643.124857  646.148601  649.172473  652.196473  655.220601  658.244857
#>  [223]  661.269241  664.293753  667.318393  670.343161  673.368057  676.393081
#>  [229]  679.418233  682.443513  685.468921  688.494457  691.520121  694.545913
#>  [235]  697.571833  700.597881  703.624057  706.650361  709.676793  712.703353
#>  [241]  715.730040  718.756856  721.783800  724.810872  727.838072  730.865400
#>  [247]  733.892856  736.920440  739.948152  742.975992  746.003960  749.032056
#>  [253]  752.060280  755.088632  758.117112  761.145720  764.174456  767.203320
#>  [259]  770.232312  773.261432  776.290680  779.320056  782.349560  785.379192
#>  [265]  788.408952  791.438840  794.468856  797.499000  800.529272  803.559672
#>  [271]  806.590200  809.620856  812.651639  815.682551  818.713591  821.744759
#>  [277]  824.776055  827.807479  830.839031  833.870711  836.902519  839.934455
#>  [283]  842.966519  845.998711  849.031031  852.063479  855.096055  858.128759
#>  [289]  861.161591  864.194551  867.227639  870.260855  873.294199  876.327671
#>  [295]  879.361271  882.394999  885.428855  888.462839  891.496951  894.531191
#>  [301]  897.565559  900.600055  903.634679  906.669431  909.704310  912.739318
#>  [307]  915.774454  918.809718  921.845110  924.880630  927.916278  930.952054
#>  [313]  933.987958  937.023990  940.060150  943.096438  946.132854  949.169398
#>  [319]  952.206070  955.242870  958.279798  961.316854  964.354038  967.391350
#>  [325]  970.428790  973.466358  976.504054  979.541878  982.579830  985.617910
#>  [331]  988.656118  991.694454  994.732918  997.771510 1000.810230 1003.849078
#>  [337] 1006.888053 1009.927157 1012.966389 1016.005749 1019.045237 1022.084853
#>  [343] 1025.124597 1028.164469 1031.204469 1034.244597 1037.284853 1040.325237
#>  [349] 1043.365749 1046.406389 1049.447157 1052.488053 1055.529077 1058.570229
#>  [355] 1061.611509 1064.652917 1067.694453 1070.736117 1073.777909 1076.819829
#>  [361] 1079.861877 1082.904053 1085.946357 1088.988789 1092.031349 1095.074037
#>  [367] 1098.116853 1101.159797 1104.202868 1107.246068 1110.289396 1113.332852
#>  [373] 1116.376436 1119.420148 1122.463988 1125.507956 1128.552052 1131.596276
#>  [379] 1134.640628 1137.685108 1140.729716 1143.774452 1146.819316 1149.864308
#>  [385] 1152.909428 1155.954676 1159.000052 1162.045556 1165.091188 1168.136948
#>  [391] 1171.182836 1174.228852 1177.274996 1180.321268 1183.367668 1186.414196
#>  [397] 1189.460852 1192.507636 1195.554548 1198.601587 1201.648755 1204.696051
#>  [403] 1207.743475 1210.791027 1213.838707 1216.886515 1219.934451 1222.982515
#>  [409] 1226.030707 1229.079027 1232.127475 1235.176051 1238.224755 1241.273587
#>  [415] 1244.322547 1247.371635 1250.420851 1253.470195 1256.519667 1259.569267
#>  [421] 1262.618995 1265.668851 1268.718835 1271.768947 1274.819187 1277.869555
#>  [427] 1280.920051 1283.970675 1287.021427 1290.072307 1293.123315 1296.174450
#>  [433] 1299.225714 1302.277106 1305.328626 1308.380274 1311.432050 1314.483954
#>  [439] 1317.535986 1320.588146 1323.640434 1326.692850 1329.745394 1332.798066
#>  [445] 1335.850866 1338.903794 1341.956850 1345.010034 1348.063346 1351.116786
#>  [451] 1354.170354 1357.224050 1360.277874 1363.331826 1366.385906 1369.440114
#>  [457] 1372.494450 1375.548914 1378.603506 1381.658226 1384.713074 1387.768050
#>  [463] 1390.823154 1393.878385 1396.933745 1399.989233 1403.044849 1406.100593
#>  [469] 1409.156465 1412.212465 1415.268593 1418.324849 1421.381233 1424.437745
#>  [475] 1427.494385 1430.551153 1433.608049 1436.665073 1439.722225 1442.779505
#>  [481] 1445.836913 1448.894449 1451.952113 1455.009905 1458.067825 1461.125873
#>  [487] 1464.184049 1467.242353 1470.300785 1473.359345 1476.418033 1479.476849
#>  [493] 1482.535793 1485.594865 1488.654065 1491.713392 1494.772848 1497.832432
#>  [499] 1500.892144 1503.951984 1507.011952 1510.072048 1513.132272 1516.192624
#>  [505] 1519.253104 1522.313712 1525.374448 1528.435312 1531.496304 1534.557424
#>  [511] 1537.618672 1540.680048 1543.741552 1546.803184 1549.864944 1552.926832
#>  [517] 1555.988848 1559.050992 1562.113264 1565.175664 1568.238192 1571.300848
#>  [523] 1574.363632 1577.426544 1580.489584 1583.552752 1586.616048 1589.679471
#>  [529] 1592.743023 1595.806703 1598.870511 1601.934447 1604.998511 1608.062703
#>  [535] 1611.127023 1614.191471 1617.256047 1620.320751 1623.385583 1626.450543
#>  [541] 1629.515631 1632.580847 1635.646191 1638.711663 1641.777263 1644.842991
#>  [547] 1647.908847 1650.974831 1654.040943 1657.107183 1660.173551 1663.240047
#>  [553] 1666.306671 1669.373423 1672.440303 1675.507311 1678.574447 1681.641711
#>  [559] 1684.709103 1687.776622 1690.844270 1693.912046 1696.979950 1700.047982
#>  [565] 1703.116142 1706.184430 1709.252846 1712.321390 1715.390062 1718.458862
#>  [571] 1721.527790 1724.596846 1727.666030 1730.735342 1733.804782 1736.874350
#>  [577] 1739.944046 1743.013870 1746.083822 1749.153902 1752.224110 1755.294446
#>  [583] 1758.364910 1761.435502 1764.506222 1767.577070 1770.648046 1773.719150
#>  [589] 1776.790382 1779.861742 1782.933230 1786.004845 1789.076589 1792.148461
#>  [595] 1795.220461 1798.292589 1801.364845 1804.437229 1807.509741 1810.582381
#>  [601] 1813.655149 1816.728045 1819.801069 1822.874221 1825.947501 1829.020909
#>  [607] 1832.094445 1835.168109 1838.241901 1841.315821 1844.389869 1847.464045
#>  [613] 1850.538349 1853.612781 1856.687341 1859.762029 1862.836845 1865.911789
#>  [619] 1868.986861 1872.062061 1875.137389 1878.212845 1881.288429 1884.364140
#>  [625] 1887.439980 1890.515948 1893.592044 1896.668268 1899.744620 1902.821100
#>  [631] 1905.897708 1908.974444 1912.051308 1915.128300 1918.205420 1921.282668
#>  [637] 1924.360044 1927.437548 1930.515180 1933.592940 1936.670828 1939.748844
#>  [643] 1942.826988 1945.905260 1948.983660 1952.062188 1955.140844 1958.219628
#>  [649] 1961.298540 1964.377580 1967.456748 1970.536044 1973.615468 1976.695020
#>  [655] 1979.774700 1982.854507 1985.934443 1989.014507 1992.094699 1995.175019
#>  [661] 1998.255467 2001.336043 2004.416747 2007.497579 2010.578539 2013.659627
#>  [667] 2016.740843 2019.822187 2022.903659 2025.985259 2029.066987 2032.148843
#>  [673] 2035.230827 2038.312939 2041.395179 2044.477547 2047.560043 2050.642667
#>  [679] 2053.725419 2056.808299 2059.891307 2062.974443 2066.057707 2069.141099
#>  [685] 2072.224619 2075.308267 2078.392042 2081.475946 2084.559978 2087.644138
#>  [691] 2090.728426 2093.812842 2096.897386 2099.982058 2103.066858 2106.151786
#>  [697] 2109.236842 2112.322026 2115.407338 2118.492778 2121.578346 2124.664042
#>  [703] 2127.749866 2130.835818 2133.921898 2137.008106 2140.094442 2143.180906
#>  [709] 2146.267498 2149.354218 2152.441066 2155.528042 2158.615146 2161.702378
#>  [715] 2164.789738 2167.877226 2170.964842 2174.052586 2177.140457 2180.228457
#>  [721] 2183.316585 2186.404841 2189.493225 2192.581737 2195.670377 2198.759145
#>  [727] 2201.848041 2204.937065 2208.026217 2211.115497 2214.204905 2217.294441
#>  [733] 2220.384105 2223.473897 2226.563817 2229.653865 2232.744041 2235.834345
#>  [739] 2238.924777 2242.015337 2245.106025 2248.196841 2251.287785 2254.378857
#>  [745] 2257.470057 2260.561385 2263.652841 2266.744425 2269.836137 2272.927977
#>  [751] 2276.019944 2279.112040 2282.204264 2285.296616 2288.389096 2291.481704
#>  [757] 2294.574440 2297.667304 2300.760296 2303.853416 2306.946664 2310.040040
#>  [763] 2313.133544 2316.227176 2319.320936 2322.414824 2325.508840 2328.602984
#>  [769] 2331.697256 2334.791656 2337.886184 2340.980840 2344.075624 2347.170536
#>  [775] 2350.265576 2353.360744 2356.456040 2359.551464 2362.647016 2365.742696
#>  [781] 2368.838504 2371.934440 2375.030503 2378.126695 2381.223015 2384.319463
#>  [787] 2387.416039 2390.512743 2393.609575 2396.706535 2399.803623 2402.900839
#>  [793] 2405.998183 2409.095655 2412.193255 2415.290983 2418.388839 2421.486823
#>  [799] 2424.584935 2427.683175 2430.781543 2433.880039 2436.978663 2440.077415
#>  [805] 2443.176295 2446.275303 2449.374439 2452.473703 2455.573095 2458.672615
#>  [811] 2461.772263 2464.872039 2467.971943 2471.071975 2474.172134 2477.272422
#>  [817] 2480.372838 2483.473382 2486.574054 2489.674854 2492.775782 2495.876838
#>  [823] 2498.978022 2502.079334 2505.180774 2508.282342 2511.384038 2514.485862
#>  [829] 2517.587814 2520.689894 2523.792102 2526.894438 2529.996902 2533.099494
#>  [835] 2536.202214 2539.305062 2542.408038 2545.511142 2548.614374 2551.717734
#>  [841] 2554.821222 2557.924838 2561.028582 2564.132454 2567.236454 2570.340581
#>  [847] 2573.444837 2576.549221 2579.653733 2582.758373 2585.863141 2588.968037
#>  [853] 2592.073061 2595.178213 2598.283493 2601.388901 2604.494437 2607.600101
#>  [859] 2610.705893 2613.811813 2616.917861 2620.024037 2623.130341 2626.236773
#>  [865] 2629.343333 2632.450021 2635.556837 2638.663781 2641.770853 2644.878053
#>  [871] 2647.985381 2651.092837 2654.200421 2657.308133 2660.415973 2663.523941
#>  [877] 2666.632037 2669.740260 2672.848612 2675.957092 2679.065700 2682.174436
#>  [883] 2685.283300 2688.392292 2691.501412 2694.610660 2697.720036 2700.829540
#>  [889] 2703.939172 2707.048932 2710.158820 2713.268836 2716.378980 2719.489252
#>  [895] 2722.599652 2725.710180 2728.820836 2731.931620 2735.042532 2738.153572
#>  [901] 2741.264740 2744.376036 2747.487460 2750.599012 2753.710692 2756.822500
#>  [907] 2759.934436 2763.046500 2766.158692 2769.271011 2772.383459 2775.496035
#>  [913] 2778.608739 2781.721571 2784.834531 2787.947619 2791.060835 2794.174179
#>  [919] 2797.287651 2800.401251 2803.514979 2806.628835 2809.742819 2812.856931
#>  [925] 2815.971171 2819.085539 2822.200035 2825.314659 2828.429411 2831.544291
#>  [931] 2834.659299 2837.774435 2840.889699 2844.005091 2847.120611 2850.236259
#>  [937] 2853.352035 2856.467939 2859.583971 2862.700131 2865.816419 2868.932834
#>  [943] 2872.049378 2875.166050 2878.282850 2881.399778 2884.516834 2887.634018
#>  [949] 2890.751330 2893.868770 2896.986338 2900.104034 2903.221858 2906.339810
#>  [955] 2909.457890 2912.576098 2915.694434 2918.812898 2921.931490 2925.050210
#>  [961] 2928.169058 2931.288034 2934.407138 2937.526370 2940.645730 2943.765218
#>  [967] 2946.884834 2950.004578 2953.124450 2956.244450 2959.364578 2962.484834
#>  [973] 2965.605218 2968.725729 2971.846369 2974.967137 2978.088033 2981.209057
#>  [979] 2984.330209 2987.451489 2990.572897 2993.694433 2996.816097 2999.937889
#>  [985] 3003.059809 3006.181857 3009.304033 3012.426337 3015.548769 3018.671329
#>  [991] 3021.794017 3024.916833 3028.039777 3031.162849 3034.286049 3037.409377
#>  [997] 3040.532833 3043.656417 3046.780129 3049.903969 3053.027937 3056.152033
#> [1003] 3059.276257 3062.400609 3065.525088 3068.649696 3071.774432 3074.899296
#> [1009] 3078.024288 3081.149408 3084.274656 3087.400032 3090.525536 3093.651168
#> [1015] 3096.776928 3099.902816 3103.028832 3106.154976 3109.281248 3112.407648
#> [1021] 3115.534176 3118.660832 3121.787616 3124.914528
#> 
#> $C341
#>    [1]   -7.004032   -4.007936   -1.011712    1.984640    4.981120    7.977728
#>    [7]   10.974464   13.971328   16.968320   19.965440   22.962688   25.960064
#>   [13]   28.957568   31.955200   34.952960   37.950848   40.948863   43.947007
#>   [19]   46.945279   49.943679   52.942207   55.940863   58.939647   61.938559
#>   [25]   64.937599   67.936767   70.936063   73.935487   76.935039   79.934719
#>   [31]   82.934527   85.934463   88.934527   91.934719   94.935039   97.935487
#>   [37]  100.936063  103.936767  106.937599  109.938559  112.939647  115.940863
#>   [43]  118.942207  121.943679  124.945279  127.947007  130.948863  133.950847
#>   [49]  136.952958  139.955198  142.957566  145.960062  148.962686  151.965438
#>   [55]  154.968318  157.971326  160.974462  163.977726  166.981118  169.984638
#>   [61]  172.988286  175.992062  178.995966  181.999998  185.004158  188.008446
#>   [67]  191.012862  194.017406  197.022078  200.026878  203.031806  206.036862
#>   [73]  209.042046  212.047358  215.052798  218.058366  221.064062  224.069886
#>   [79]  227.075838  230.081918  233.088125  236.094461  239.100925  242.107517
#>   [85]  245.114237  248.121085  251.128061  254.135165  257.142397  260.149757
#>   [91]  263.157245  266.164861  269.172605  272.180477  275.188477  278.196605
#>   [97]  281.204861  284.213245  287.221757  290.230397  293.239165  296.248061
#>  [103]  299.257085  302.266237  305.275517  308.284925  311.294461  314.304125
#>  [109]  317.313917  320.323837  323.333885  326.344061  329.354364  332.364796
#>  [115]  335.375356  338.386044  341.396860  344.407804  347.418876  350.430076
#>  [121]  353.441404  356.452860  359.464444  362.476156  365.487996  368.499964
#>  [127]  371.512060  374.524284  377.536636  380.549116  383.561724  386.574460
#>  [133]  389.587324  392.600316  395.613436  398.626684  401.640060  404.653564
#>  [139]  407.667196  410.680956  413.694844  416.708860  419.723004  422.737276
#>  [145]  425.751675  428.766203  431.780859  434.795643  437.810555  440.825595
#>  [151]  443.840763  446.856059  449.871483  452.887035  455.902715  458.918523
#>  [157]  461.934459  464.950523  467.966715  470.983035  473.999483  477.016059
#>  [163]  480.032763  483.049595  486.066555  489.083643  492.100859  495.118203
#>  [169]  498.135675  501.153275  504.171003  507.188859  510.206843  513.224955
#>  [175]  516.243195  519.261563  522.280058  525.298682  528.317434  531.336314
#>  [181]  534.355322  537.374458  540.393722  543.413114  546.432634  549.452282
#>  [187]  552.472058  555.491962  558.511994  561.532154  564.552442  567.572858
#>  [193]  570.593402  573.614074  576.634874  579.655802  582.676858  585.698042
#>  [199]  588.719354  591.740794  594.762362  597.784058  600.805882  603.827834
#>  [205]  606.849914  609.872122  612.894458  615.916922  618.939513  621.962233
#>  [211]  624.985081  628.008057  631.031161  634.054393  637.077753  640.101241
#>  [217]  643.124857  646.148601  649.172473  652.196473  655.220601  658.244857
#>  [223]  661.269241  664.293753  667.318393  670.343161  673.368057  676.393081
#>  [229]  679.418233  682.443513  685.468921  688.494457  691.520121  694.545913
#>  [235]  697.571833  700.597881  703.624057  706.650361  709.676793  712.703353
#>  [241]  715.730040  718.756856  721.783800  724.810872  727.838072  730.865400
#>  [247]  733.892856  736.920440  739.948152  742.975992  746.003960  749.032056
#>  [253]  752.060280  755.088632  758.117112  761.145720  764.174456  767.203320
#>  [259]  770.232312  773.261432  776.290680  779.320056  782.349560  785.379192
#>  [265]  788.408952  791.438840  794.468856  797.499000  800.529272  803.559672
#>  [271]  806.590200  809.620856  812.651639  815.682551  818.713591  821.744759
#>  [277]  824.776055  827.807479  830.839031  833.870711  836.902519  839.934455
#>  [283]  842.966519  845.998711  849.031031  852.063479  855.096055  858.128759
#>  [289]  861.161591  864.194551  867.227639  870.260855  873.294199  876.327671
#>  [295]  879.361271  882.394999  885.428855  888.462839  891.496951  894.531191
#>  [301]  897.565559  900.600055  903.634679  906.669431  909.704310  912.739318
#>  [307]  915.774454  918.809718  921.845110  924.880630  927.916278  930.952054
#>  [313]  933.987958  937.023990  940.060150  943.096438  946.132854  949.169398
#>  [319]  952.206070  955.242870  958.279798  961.316854  964.354038  967.391350
#>  [325]  970.428790  973.466358  976.504054  979.541878  982.579830  985.617910
#>  [331]  988.656118  991.694454  994.732918  997.771510 1000.810230 1003.849078
#>  [337] 1006.888053 1009.927157 1012.966389 1016.005749 1019.045237 1022.084853
#>  [343] 1025.124597 1028.164469 1031.204469 1034.244597 1037.284853 1040.325237
#>  [349] 1043.365749 1046.406389 1049.447157 1052.488053 1055.529077 1058.570229
#>  [355] 1061.611509 1064.652917 1067.694453 1070.736117 1073.777909 1076.819829
#>  [361] 1079.861877 1082.904053 1085.946357 1088.988789 1092.031349 1095.074037
#>  [367] 1098.116853 1101.159797 1104.202868 1107.246068 1110.289396 1113.332852
#>  [373] 1116.376436 1119.420148 1122.463988 1125.507956 1128.552052 1131.596276
#>  [379] 1134.640628 1137.685108 1140.729716 1143.774452 1146.819316 1149.864308
#>  [385] 1152.909428 1155.954676 1159.000052 1162.045556 1165.091188 1168.136948
#>  [391] 1171.182836 1174.228852 1177.274996 1180.321268 1183.367668 1186.414196
#>  [397] 1189.460852 1192.507636 1195.554548 1198.601587 1201.648755 1204.696051
#>  [403] 1207.743475 1210.791027 1213.838707 1216.886515 1219.934451 1222.982515
#>  [409] 1226.030707 1229.079027 1232.127475 1235.176051 1238.224755 1241.273587
#>  [415] 1244.322547 1247.371635 1250.420851 1253.470195 1256.519667 1259.569267
#>  [421] 1262.618995 1265.668851 1268.718835 1271.768947 1274.819187 1277.869555
#>  [427] 1280.920051 1283.970675 1287.021427 1290.072307 1293.123315 1296.174450
#>  [433] 1299.225714 1302.277106 1305.328626 1308.380274 1311.432050 1314.483954
#>  [439] 1317.535986 1320.588146 1323.640434 1326.692850 1329.745394 1332.798066
#>  [445] 1335.850866 1338.903794 1341.956850 1345.010034 1348.063346 1351.116786
#>  [451] 1354.170354 1357.224050 1360.277874 1363.331826 1366.385906 1369.440114
#>  [457] 1372.494450 1375.548914 1378.603506 1381.658226 1384.713074 1387.768050
#>  [463] 1390.823154 1393.878385 1396.933745 1399.989233 1403.044849 1406.100593
#>  [469] 1409.156465 1412.212465 1415.268593 1418.324849 1421.381233 1424.437745
#>  [475] 1427.494385 1430.551153 1433.608049 1436.665073 1439.722225 1442.779505
#>  [481] 1445.836913 1448.894449 1451.952113 1455.009905 1458.067825 1461.125873
#>  [487] 1464.184049 1467.242353 1470.300785 1473.359345 1476.418033 1479.476849
#>  [493] 1482.535793 1485.594865 1488.654065 1491.713392 1494.772848 1497.832432
#>  [499] 1500.892144 1503.951984 1507.011952 1510.072048 1513.132272 1516.192624
#>  [505] 1519.253104 1522.313712 1525.374448 1528.435312 1531.496304 1534.557424
#>  [511] 1537.618672 1540.680048 1543.741552 1546.803184 1549.864944 1552.926832
#>  [517] 1555.988848 1559.050992 1562.113264 1565.175664 1568.238192 1571.300848
#>  [523] 1574.363632 1577.426544 1580.489584 1583.552752 1586.616048 1589.679471
#>  [529] 1592.743023 1595.806703 1598.870511 1601.934447 1604.998511 1608.062703
#>  [535] 1611.127023 1614.191471 1617.256047 1620.320751 1623.385583 1626.450543
#>  [541] 1629.515631 1632.580847 1635.646191 1638.711663 1641.777263 1644.842991
#>  [547] 1647.908847 1650.974831 1654.040943 1657.107183 1660.173551 1663.240047
#>  [553] 1666.306671 1669.373423 1672.440303 1675.507311 1678.574447 1681.641711
#>  [559] 1684.709103 1687.776622 1690.844270 1693.912046 1696.979950 1700.047982
#>  [565] 1703.116142 1706.184430 1709.252846 1712.321390 1715.390062 1718.458862
#>  [571] 1721.527790 1724.596846 1727.666030 1730.735342 1733.804782 1736.874350
#>  [577] 1739.944046 1743.013870 1746.083822 1749.153902 1752.224110 1755.294446
#>  [583] 1758.364910 1761.435502 1764.506222 1767.577070 1770.648046 1773.719150
#>  [589] 1776.790382 1779.861742 1782.933230 1786.004845 1789.076589 1792.148461
#>  [595] 1795.220461 1798.292589 1801.364845 1804.437229 1807.509741 1810.582381
#>  [601] 1813.655149 1816.728045 1819.801069 1822.874221 1825.947501 1829.020909
#>  [607] 1832.094445 1835.168109 1838.241901 1841.315821 1844.389869 1847.464045
#>  [613] 1850.538349 1853.612781 1856.687341 1859.762029 1862.836845 1865.911789
#>  [619] 1868.986861 1872.062061 1875.137389 1878.212845 1881.288429 1884.364140
#>  [625] 1887.439980 1890.515948 1893.592044 1896.668268 1899.744620 1902.821100
#>  [631] 1905.897708 1908.974444 1912.051308 1915.128300 1918.205420 1921.282668
#>  [637] 1924.360044 1927.437548 1930.515180 1933.592940 1936.670828 1939.748844
#>  [643] 1942.826988 1945.905260 1948.983660 1952.062188 1955.140844 1958.219628
#>  [649] 1961.298540 1964.377580 1967.456748 1970.536044 1973.615468 1976.695020
#>  [655] 1979.774700 1982.854507 1985.934443 1989.014507 1992.094699 1995.175019
#>  [661] 1998.255467 2001.336043 2004.416747 2007.497579 2010.578539 2013.659627
#>  [667] 2016.740843 2019.822187 2022.903659 2025.985259 2029.066987 2032.148843
#>  [673] 2035.230827 2038.312939 2041.395179 2044.477547 2047.560043 2050.642667
#>  [679] 2053.725419 2056.808299 2059.891307 2062.974443 2066.057707 2069.141099
#>  [685] 2072.224619 2075.308267 2078.392042 2081.475946 2084.559978 2087.644138
#>  [691] 2090.728426 2093.812842 2096.897386 2099.982058 2103.066858 2106.151786
#>  [697] 2109.236842 2112.322026 2115.407338 2118.492778 2121.578346 2124.664042
#>  [703] 2127.749866 2130.835818 2133.921898 2137.008106 2140.094442 2143.180906
#>  [709] 2146.267498 2149.354218 2152.441066 2155.528042 2158.615146 2161.702378
#>  [715] 2164.789738 2167.877226 2170.964842 2174.052586 2177.140457 2180.228457
#>  [721] 2183.316585 2186.404841 2189.493225 2192.581737 2195.670377 2198.759145
#>  [727] 2201.848041 2204.937065 2208.026217 2211.115497 2214.204905 2217.294441
#>  [733] 2220.384105 2223.473897 2226.563817 2229.653865 2232.744041 2235.834345
#>  [739] 2238.924777 2242.015337 2245.106025 2248.196841 2251.287785 2254.378857
#>  [745] 2257.470057 2260.561385 2263.652841 2266.744425 2269.836137 2272.927977
#>  [751] 2276.019944 2279.112040 2282.204264 2285.296616 2288.389096 2291.481704
#>  [757] 2294.574440 2297.667304 2300.760296 2303.853416 2306.946664 2310.040040
#>  [763] 2313.133544 2316.227176 2319.320936 2322.414824 2325.508840 2328.602984
#>  [769] 2331.697256 2334.791656 2337.886184 2340.980840 2344.075624 2347.170536
#>  [775] 2350.265576 2353.360744 2356.456040 2359.551464 2362.647016 2365.742696
#>  [781] 2368.838504 2371.934440 2375.030503 2378.126695 2381.223015 2384.319463
#>  [787] 2387.416039 2390.512743 2393.609575 2396.706535 2399.803623 2402.900839
#>  [793] 2405.998183 2409.095655 2412.193255 2415.290983 2418.388839 2421.486823
#>  [799] 2424.584935 2427.683175 2430.781543 2433.880039 2436.978663 2440.077415
#>  [805] 2443.176295 2446.275303 2449.374439 2452.473703 2455.573095 2458.672615
#>  [811] 2461.772263 2464.872039 2467.971943 2471.071975 2474.172134 2477.272422
#>  [817] 2480.372838 2483.473382 2486.574054 2489.674854 2492.775782 2495.876838
#>  [823] 2498.978022 2502.079334 2505.180774 2508.282342 2511.384038 2514.485862
#>  [829] 2517.587814 2520.689894 2523.792102 2526.894438 2529.996902 2533.099494
#>  [835] 2536.202214 2539.305062 2542.408038 2545.511142 2548.614374 2551.717734
#>  [841] 2554.821222 2557.924838 2561.028582 2564.132454 2567.236454 2570.340581
#>  [847] 2573.444837 2576.549221 2579.653733 2582.758373 2585.863141 2588.968037
#>  [853] 2592.073061 2595.178213 2598.283493 2601.388901 2604.494437 2607.600101
#>  [859] 2610.705893 2613.811813 2616.917861 2620.024037 2623.130341 2626.236773
#>  [865] 2629.343333 2632.450021 2635.556837 2638.663781 2641.770853 2644.878053
#>  [871] 2647.985381 2651.092837 2654.200421 2657.308133 2660.415973 2663.523941
#>  [877] 2666.632037 2669.740260 2672.848612 2675.957092 2679.065700 2682.174436
#>  [883] 2685.283300 2688.392292 2691.501412 2694.610660 2697.720036 2700.829540
#>  [889] 2703.939172 2707.048932 2710.158820 2713.268836 2716.378980 2719.489252
#>  [895] 2722.599652 2725.710180 2728.820836 2731.931620 2735.042532 2738.153572
#>  [901] 2741.264740 2744.376036 2747.487460 2750.599012 2753.710692 2756.822500
#>  [907] 2759.934436 2763.046500 2766.158692 2769.271011 2772.383459 2775.496035
#>  [913] 2778.608739 2781.721571 2784.834531 2787.947619 2791.060835 2794.174179
#>  [919] 2797.287651 2800.401251 2803.514979 2806.628835 2809.742819 2812.856931
#>  [925] 2815.971171 2819.085539 2822.200035 2825.314659 2828.429411 2831.544291
#>  [931] 2834.659299 2837.774435 2840.889699 2844.005091 2847.120611 2850.236259
#>  [937] 2853.352035 2856.467939 2859.583971 2862.700131 2865.816419 2868.932834
#>  [943] 2872.049378 2875.166050 2878.282850 2881.399778 2884.516834 2887.634018
#>  [949] 2890.751330 2893.868770 2896.986338 2900.104034 2903.221858 2906.339810
#>  [955] 2909.457890 2912.576098 2915.694434 2918.812898 2921.931490 2925.050210
#>  [961] 2928.169058 2931.288034 2934.407138 2937.526370 2940.645730 2943.765218
#>  [967] 2946.884834 2950.004578 2953.124450 2956.244450 2959.364578 2962.484834
#>  [973] 2965.605218 2968.725729 2971.846369 2974.967137 2978.088033 2981.209057
#>  [979] 2984.330209 2987.451489 2990.572897 2993.694433 2996.816097 2999.937889
#>  [985] 3003.059809 3006.181857 3009.304033 3012.426337 3015.548769 3018.671329
#>  [991] 3021.794017 3024.916833 3028.039777 3031.162849 3034.286049 3037.409377
#>  [997] 3040.532833 3043.656417 3046.780129 3049.903969 3053.027937 3056.152033
#> [1003] 3059.276257 3062.400609 3065.525088 3068.649696 3071.774432 3074.899296
#> [1009] 3078.024288 3081.149408 3084.274656 3087.400032 3090.525536 3093.651168
#> [1015] 3096.776928 3099.902816 3103.028832 3106.154976 3109.281248 3112.407648
#> [1021] 3115.534176 3118.660832 3121.787616 3124.914528
#> 
#> $C347
#>    [1]   -7.004032   -4.007936   -1.011712    1.984640    4.981120    7.977728
#>    [7]   10.974464   13.971328   16.968320   19.965440   22.962688   25.960064
#>   [13]   28.957568   31.955200   34.952960   37.950848   40.948863   43.947007
#>   [19]   46.945279   49.943679   52.942207   55.940863   58.939647   61.938559
#>   [25]   64.937599   67.936767   70.936063   73.935487   76.935039   79.934719
#>   [31]   82.934527   85.934463   88.934527   91.934719   94.935039   97.935487
#>   [37]  100.936063  103.936767  106.937599  109.938559  112.939647  115.940863
#>   [43]  118.942207  121.943679  124.945279  127.947007  130.948863  133.950847
#>   [49]  136.952958  139.955198  142.957566  145.960062  148.962686  151.965438
#>   [55]  154.968318  157.971326  160.974462  163.977726  166.981118  169.984638
#>   [61]  172.988286  175.992062  178.995966  181.999998  185.004158  188.008446
#>   [67]  191.012862  194.017406  197.022078  200.026878  203.031806  206.036862
#>   [73]  209.042046  212.047358  215.052798  218.058366  221.064062  224.069886
#>   [79]  227.075838  230.081918  233.088125  236.094461  239.100925  242.107517
#>   [85]  245.114237  248.121085  251.128061  254.135165  257.142397  260.149757
#>   [91]  263.157245  266.164861  269.172605  272.180477  275.188477  278.196605
#>   [97]  281.204861  284.213245  287.221757  290.230397  293.239165  296.248061
#>  [103]  299.257085  302.266237  305.275517  308.284925  311.294461  314.304125
#>  [109]  317.313917  320.323837  323.333885  326.344061  329.354364  332.364796
#>  [115]  335.375356  338.386044  341.396860  344.407804  347.418876  350.430076
#>  [121]  353.441404  356.452860  359.464444  362.476156  365.487996  368.499964
#>  [127]  371.512060  374.524284  377.536636  380.549116  383.561724  386.574460
#>  [133]  389.587324  392.600316  395.613436  398.626684  401.640060  404.653564
#>  [139]  407.667196  410.680956  413.694844  416.708860  419.723004  422.737276
#>  [145]  425.751675  428.766203  431.780859  434.795643  437.810555  440.825595
#>  [151]  443.840763  446.856059  449.871483  452.887035  455.902715  458.918523
#>  [157]  461.934459  464.950523  467.966715  470.983035  473.999483  477.016059
#>  [163]  480.032763  483.049595  486.066555  489.083643  492.100859  495.118203
#>  [169]  498.135675  501.153275  504.171003  507.188859  510.206843  513.224955
#>  [175]  516.243195  519.261563  522.280058  525.298682  528.317434  531.336314
#>  [181]  534.355322  537.374458  540.393722  543.413114  546.432634  549.452282
#>  [187]  552.472058  555.491962  558.511994  561.532154  564.552442  567.572858
#>  [193]  570.593402  573.614074  576.634874  579.655802  582.676858  585.698042
#>  [199]  588.719354  591.740794  594.762362  597.784058  600.805882  603.827834
#>  [205]  606.849914  609.872122  612.894458  615.916922  618.939513  621.962233
#>  [211]  624.985081  628.008057  631.031161  634.054393  637.077753  640.101241
#>  [217]  643.124857  646.148601  649.172473  652.196473  655.220601  658.244857
#>  [223]  661.269241  664.293753  667.318393  670.343161  673.368057  676.393081
#>  [229]  679.418233  682.443513  685.468921  688.494457  691.520121  694.545913
#>  [235]  697.571833  700.597881  703.624057  706.650361  709.676793  712.703353
#>  [241]  715.730040  718.756856  721.783800  724.810872  727.838072  730.865400
#>  [247]  733.892856  736.920440  739.948152  742.975992  746.003960  749.032056
#>  [253]  752.060280  755.088632  758.117112  761.145720  764.174456  767.203320
#>  [259]  770.232312  773.261432  776.290680  779.320056  782.349560  785.379192
#>  [265]  788.408952  791.438840  794.468856  797.499000  800.529272  803.559672
#>  [271]  806.590200  809.620856  812.651639  815.682551  818.713591  821.744759
#>  [277]  824.776055  827.807479  830.839031  833.870711  836.902519  839.934455
#>  [283]  842.966519  845.998711  849.031031  852.063479  855.096055  858.128759
#>  [289]  861.161591  864.194551  867.227639  870.260855  873.294199  876.327671
#>  [295]  879.361271  882.394999  885.428855  888.462839  891.496951  894.531191
#>  [301]  897.565559  900.600055  903.634679  906.669431  909.704310  912.739318
#>  [307]  915.774454  918.809718  921.845110  924.880630  927.916278  930.952054
#>  [313]  933.987958  937.023990  940.060150  943.096438  946.132854  949.169398
#>  [319]  952.206070  955.242870  958.279798  961.316854  964.354038  967.391350
#>  [325]  970.428790  973.466358  976.504054  979.541878  982.579830  985.617910
#>  [331]  988.656118  991.694454  994.732918  997.771510 1000.810230 1003.849078
#>  [337] 1006.888053 1009.927157 1012.966389 1016.005749 1019.045237 1022.084853
#>  [343] 1025.124597 1028.164469 1031.204469 1034.244597 1037.284853 1040.325237
#>  [349] 1043.365749 1046.406389 1049.447157 1052.488053 1055.529077 1058.570229
#>  [355] 1061.611509 1064.652917 1067.694453 1070.736117 1073.777909 1076.819829
#>  [361] 1079.861877 1082.904053 1085.946357 1088.988789 1092.031349 1095.074037
#>  [367] 1098.116853 1101.159797 1104.202868 1107.246068 1110.289396 1113.332852
#>  [373] 1116.376436 1119.420148 1122.463988 1125.507956 1128.552052 1131.596276
#>  [379] 1134.640628 1137.685108 1140.729716 1143.774452 1146.819316 1149.864308
#>  [385] 1152.909428 1155.954676 1159.000052 1162.045556 1165.091188 1168.136948
#>  [391] 1171.182836 1174.228852 1177.274996 1180.321268 1183.367668 1186.414196
#>  [397] 1189.460852 1192.507636 1195.554548 1198.601587 1201.648755 1204.696051
#>  [403] 1207.743475 1210.791027 1213.838707 1216.886515 1219.934451 1222.982515
#>  [409] 1226.030707 1229.079027 1232.127475 1235.176051 1238.224755 1241.273587
#>  [415] 1244.322547 1247.371635 1250.420851 1253.470195 1256.519667 1259.569267
#>  [421] 1262.618995 1265.668851 1268.718835 1271.768947 1274.819187 1277.869555
#>  [427] 1280.920051 1283.970675 1287.021427 1290.072307 1293.123315 1296.174450
#>  [433] 1299.225714 1302.277106 1305.328626 1308.380274 1311.432050 1314.483954
#>  [439] 1317.535986 1320.588146 1323.640434 1326.692850 1329.745394 1332.798066
#>  [445] 1335.850866 1338.903794 1341.956850 1345.010034 1348.063346 1351.116786
#>  [451] 1354.170354 1357.224050 1360.277874 1363.331826 1366.385906 1369.440114
#>  [457] 1372.494450 1375.548914 1378.603506 1381.658226 1384.713074 1387.768050
#>  [463] 1390.823154 1393.878385 1396.933745 1399.989233 1403.044849 1406.100593
#>  [469] 1409.156465 1412.212465 1415.268593 1418.324849 1421.381233 1424.437745
#>  [475] 1427.494385 1430.551153 1433.608049 1436.665073 1439.722225 1442.779505
#>  [481] 1445.836913 1448.894449 1451.952113 1455.009905 1458.067825 1461.125873
#>  [487] 1464.184049 1467.242353 1470.300785 1473.359345 1476.418033 1479.476849
#>  [493] 1482.535793 1485.594865 1488.654065 1491.713392 1494.772848 1497.832432
#>  [499] 1500.892144 1503.951984 1507.011952 1510.072048 1513.132272 1516.192624
#>  [505] 1519.253104 1522.313712 1525.374448 1528.435312 1531.496304 1534.557424
#>  [511] 1537.618672 1540.680048 1543.741552 1546.803184 1549.864944 1552.926832
#>  [517] 1555.988848 1559.050992 1562.113264 1565.175664 1568.238192 1571.300848
#>  [523] 1574.363632 1577.426544 1580.489584 1583.552752 1586.616048 1589.679471
#>  [529] 1592.743023 1595.806703 1598.870511 1601.934447 1604.998511 1608.062703
#>  [535] 1611.127023 1614.191471 1617.256047 1620.320751 1623.385583 1626.450543
#>  [541] 1629.515631 1632.580847 1635.646191 1638.711663 1641.777263 1644.842991
#>  [547] 1647.908847 1650.974831 1654.040943 1657.107183 1660.173551 1663.240047
#>  [553] 1666.306671 1669.373423 1672.440303 1675.507311 1678.574447 1681.641711
#>  [559] 1684.709103 1687.776622 1690.844270 1693.912046 1696.979950 1700.047982
#>  [565] 1703.116142 1706.184430 1709.252846 1712.321390 1715.390062 1718.458862
#>  [571] 1721.527790 1724.596846 1727.666030 1730.735342 1733.804782 1736.874350
#>  [577] 1739.944046 1743.013870 1746.083822 1749.153902 1752.224110 1755.294446
#>  [583] 1758.364910 1761.435502 1764.506222 1767.577070 1770.648046 1773.719150
#>  [589] 1776.790382 1779.861742 1782.933230 1786.004845 1789.076589 1792.148461
#>  [595] 1795.220461 1798.292589 1801.364845 1804.437229 1807.509741 1810.582381
#>  [601] 1813.655149 1816.728045 1819.801069 1822.874221 1825.947501 1829.020909
#>  [607] 1832.094445 1835.168109 1838.241901 1841.315821 1844.389869 1847.464045
#>  [613] 1850.538349 1853.612781 1856.687341 1859.762029 1862.836845 1865.911789
#>  [619] 1868.986861 1872.062061 1875.137389 1878.212845 1881.288429 1884.364140
#>  [625] 1887.439980 1890.515948 1893.592044 1896.668268 1899.744620 1902.821100
#>  [631] 1905.897708 1908.974444 1912.051308 1915.128300 1918.205420 1921.282668
#>  [637] 1924.360044 1927.437548 1930.515180 1933.592940 1936.670828 1939.748844
#>  [643] 1942.826988 1945.905260 1948.983660 1952.062188 1955.140844 1958.219628
#>  [649] 1961.298540 1964.377580 1967.456748 1970.536044 1973.615468 1976.695020
#>  [655] 1979.774700 1982.854507 1985.934443 1989.014507 1992.094699 1995.175019
#>  [661] 1998.255467 2001.336043 2004.416747 2007.497579 2010.578539 2013.659627
#>  [667] 2016.740843 2019.822187 2022.903659 2025.985259 2029.066987 2032.148843
#>  [673] 2035.230827 2038.312939 2041.395179 2044.477547 2047.560043 2050.642667
#>  [679] 2053.725419 2056.808299 2059.891307 2062.974443 2066.057707 2069.141099
#>  [685] 2072.224619 2075.308267 2078.392042 2081.475946 2084.559978 2087.644138
#>  [691] 2090.728426 2093.812842 2096.897386 2099.982058 2103.066858 2106.151786
#>  [697] 2109.236842 2112.322026 2115.407338 2118.492778 2121.578346 2124.664042
#>  [703] 2127.749866 2130.835818 2133.921898 2137.008106 2140.094442 2143.180906
#>  [709] 2146.267498 2149.354218 2152.441066 2155.528042 2158.615146 2161.702378
#>  [715] 2164.789738 2167.877226 2170.964842 2174.052586 2177.140457 2180.228457
#>  [721] 2183.316585 2186.404841 2189.493225 2192.581737 2195.670377 2198.759145
#>  [727] 2201.848041 2204.937065 2208.026217 2211.115497 2214.204905 2217.294441
#>  [733] 2220.384105 2223.473897 2226.563817 2229.653865 2232.744041 2235.834345
#>  [739] 2238.924777 2242.015337 2245.106025 2248.196841 2251.287785 2254.378857
#>  [745] 2257.470057 2260.561385 2263.652841 2266.744425 2269.836137 2272.927977
#>  [751] 2276.019944 2279.112040 2282.204264 2285.296616 2288.389096 2291.481704
#>  [757] 2294.574440 2297.667304 2300.760296 2303.853416 2306.946664 2310.040040
#>  [763] 2313.133544 2316.227176 2319.320936 2322.414824 2325.508840 2328.602984
#>  [769] 2331.697256 2334.791656 2337.886184 2340.980840 2344.075624 2347.170536
#>  [775] 2350.265576 2353.360744 2356.456040 2359.551464 2362.647016 2365.742696
#>  [781] 2368.838504 2371.934440 2375.030503 2378.126695 2381.223015 2384.319463
#>  [787] 2387.416039 2390.512743 2393.609575 2396.706535 2399.803623 2402.900839
#>  [793] 2405.998183 2409.095655 2412.193255 2415.290983 2418.388839 2421.486823
#>  [799] 2424.584935 2427.683175 2430.781543 2433.880039 2436.978663 2440.077415
#>  [805] 2443.176295 2446.275303 2449.374439 2452.473703 2455.573095 2458.672615
#>  [811] 2461.772263 2464.872039 2467.971943 2471.071975 2474.172134 2477.272422
#>  [817] 2480.372838 2483.473382 2486.574054 2489.674854 2492.775782 2495.876838
#>  [823] 2498.978022 2502.079334 2505.180774 2508.282342 2511.384038 2514.485862
#>  [829] 2517.587814 2520.689894 2523.792102 2526.894438 2529.996902 2533.099494
#>  [835] 2536.202214 2539.305062 2542.408038 2545.511142 2548.614374 2551.717734
#>  [841] 2554.821222 2557.924838 2561.028582 2564.132454 2567.236454 2570.340581
#>  [847] 2573.444837 2576.549221 2579.653733 2582.758373 2585.863141 2588.968037
#>  [853] 2592.073061 2595.178213 2598.283493 2601.388901 2604.494437 2607.600101
#>  [859] 2610.705893 2613.811813 2616.917861 2620.024037 2623.130341 2626.236773
#>  [865] 2629.343333 2632.450021 2635.556837 2638.663781 2641.770853 2644.878053
#>  [871] 2647.985381 2651.092837 2654.200421 2657.308133 2660.415973 2663.523941
#>  [877] 2666.632037 2669.740260 2672.848612 2675.957092 2679.065700 2682.174436
#>  [883] 2685.283300 2688.392292 2691.501412 2694.610660 2697.720036 2700.829540
#>  [889] 2703.939172 2707.048932 2710.158820 2713.268836 2716.378980 2719.489252
#>  [895] 2722.599652 2725.710180 2728.820836 2731.931620 2735.042532 2738.153572
#>  [901] 2741.264740 2744.376036 2747.487460 2750.599012 2753.710692 2756.822500
#>  [907] 2759.934436 2763.046500 2766.158692 2769.271011 2772.383459 2775.496035
#>  [913] 2778.608739 2781.721571 2784.834531 2787.947619 2791.060835 2794.174179
#>  [919] 2797.287651 2800.401251 2803.514979 2806.628835 2809.742819 2812.856931
#>  [925] 2815.971171 2819.085539 2822.200035 2825.314659 2828.429411 2831.544291
#>  [931] 2834.659299 2837.774435 2840.889699 2844.005091 2847.120611 2850.236259
#>  [937] 2853.352035 2856.467939 2859.583971 2862.700131 2865.816419 2868.932834
#>  [943] 2872.049378 2875.166050 2878.282850 2881.399778 2884.516834 2887.634018
#>  [949] 2890.751330 2893.868770 2896.986338 2900.104034 2903.221858 2906.339810
#>  [955] 2909.457890 2912.576098 2915.694434 2918.812898 2921.931490 2925.050210
#>  [961] 2928.169058 2931.288034 2934.407138 2937.526370 2940.645730 2943.765218
#>  [967] 2946.884834 2950.004578 2953.124450 2956.244450 2959.364578 2962.484834
#>  [973] 2965.605218 2968.725729 2971.846369 2974.967137 2978.088033 2981.209057
#>  [979] 2984.330209 2987.451489 2990.572897 2993.694433 2996.816097 2999.937889
#>  [985] 3003.059809 3006.181857 3009.304033 3012.426337 3015.548769 3018.671329
#>  [991] 3021.794017 3024.916833 3028.039777 3031.162849 3034.286049 3037.409377
#>  [997] 3040.532833 3043.656417 3046.780129 3049.903969 3053.027937 3056.152033
#> [1003] 3059.276257 3062.400609 3065.525088 3068.649696 3071.774432 3074.899296
#> [1009] 3078.024288 3081.149408 3084.274656 3087.400032 3090.525536 3093.651168
#> [1015] 3096.776928 3099.902816 3103.028832 3106.154976 3109.281248 3112.407648
#> [1021] 3115.534176 3118.660832 3121.787616 3124.914528
#> 

spectra[[1]]

#> Gamma spectrum:
#>   Reference: BRIQUE
#>   Instrument: InSpector 1000
#>   Date: 2019-03-27
#>   Number of chanels: 1024
#>   Energy range (keV): -7-3124.91

spectra[["BRIQUE"]]

#> Gamma spectrum:
#>   Reference: BRIQUE
#>   Instrument: InSpector 1000
#>   Date: 2019-03-27
#>   Number of chanels: 1024
#>   Energy range (keV): -7-3124.91