An S4 Class to Represent a Gamma Sectrum

Represents a single spectrum of a gamma ray spectrometry measurement.

Slots

hash

A character string giving the 32-byte MD5 hash of the imported file.

name

A character string the measurement reference.

date

A POSIXct element giving the measurement date and time.

instrument

A character string giving the instrument name.

file_format

A character string giving the format of the imported file.

live_time

A numeric value.

real_time

A numeric value.

chanel

A integer vector giving the channel number. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero).

energy

A numeric vector giving the gamma ray's energy (in keV).

count

A numeric vector giving the counts number for each channel. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero).

rate

A numeric vector the count rate (in 1/s) for each channel.

calibration

A linear model used for energy scale calibration (see calibrate_energy).

dose_rate

A length-two numeric vector giving the dose rate and corresponding error.

Note

This class retains copy construction.

Access

In the code snippets below, x is a GammaSpectrum object.

get_hash(x)

Get the MD5 hash of the raw data file.

get_chanels(x)

Get the number of chanels in x.

get_energy(x)

Get the energy range of x.

get_dose(x)

Get the dose rate of x.

Coerce

In the code snippets below, x is a GammaSpectrum object.

as(x, "matrix")

Coerces x to a matrix.

as(x, "data.frame")

Coerces x to a data frame.

Subset

In the code snippets below, x is a GammaSpectrum object.

x[[i]]

Extracts informations from a slot selected by subscript i. i is a character vector of length one and will be matched to the name of the slots.

See also

Examples

# Import a Canberra CNF file
spc_file <- system.file("extdata/test_CNF.cnf", package = "gamma")
(spectrum <- read(spc_file))

#> Gamma spectrum:
#> *  name: test_CNF
#> *  date: 2019-02-07 11:48:18
#> *  live time: 3385.54
#> *  real time: 3403.67
#> *  chanels: 1024
#> *  energy range: -7 3124.91

## Access
get_hash(spectrum)

#> [1] "e0a2c67173cf7f407db7148ae0058bbf"

get_chanels(spectrum)

#> [1] 1024

get_energy(spectrum)

#> [1]   -7.004032 3124.914528

get_dose(spectrum)

#> [1] 0 0

# Coerce
df <- as(spectrum, "data.frame")
head(df)

#>   chanel    energy count rate
#> 1      1 -7.004032     0    0
#> 2      2 -4.007936     0    0
#> 3      3 -1.011712     0    0
#> 4      4  1.984640     0    0
#> 5      5  4.981120     0    0
#> 6      6  7.977728     0    0

tail(df)

#>      chanel   energy count         rate
#> 1019   1019 3109.281     1 0.0002953739
#> 1020   1020 3112.408     1 0.0002953739
#> 1021   1021 3115.534     2 0.0005907477
#> 1022   1022 3118.661     0 0.0000000000
#> 1023   1023 3121.788     1 0.0002953739
#> 1024   1024 3124.915     0 0.0000000000

# Subset
spectrum[["hash"]]

#> [1] "e0a2c67173cf7f407db7148ae0058bbf"

spectrum[["name"]]

#> [1] "test_CNF"

spectrum[["date"]]

#> [1] "2019-02-07 11:48:18 UTC"

spectrum[["instrument"]]

#> [1] "InSpector 1000"

spectrum[["file_format"]]

#> [1] "CNF"

spectrum[["live_time"]]

#> [1] 3385.54

spectrum[["real_time"]]

#> [1] 3403.67

spectrum[["calibration"]]

#> 
#> Call:
#> stats::lm(formula = 0 ~ 0)
#> 
#> No coefficients
#>