gamma: Dose Rate Estimation from in-Situ Gamma-Ray Spectrometry Measurements

Process in-situ Gamma-Ray Spectrometry for Luminescence Dating. This package allows to import, inspect and correct the energy shifts of Gamma-ray spectra. It provides methods for estimating the gamma dose rate by the use of a calibration curve. The package only supports Canberra CNF and TKA files.

Details

Package:	gamma

Type:	Package

Version:	1.0.1.9000

License:	GPL-3

Zenodo:	doi:10.5281/zenodo.2652393

Note

This work received a state financial support managed by the Agence Nationale de la Recherche (France) through the program Investissements d'avenir (ref. ANR-10-LABX-52 and 11-IDEX-0001).

See also

Author

Full list of authors and contributors (alphabetic order)

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Christelle Lahaye	Université Bordeaux Montaigne, France

Brice Lebrun	Université Bordeaux Montaigne, France

Guilhem Paradol	Commissariat à l'Énergie Atomique, France

Magali Rizza	Aix-Marseille Université, France

Package maintainer

Brice Lebrun
brice.lebrun@u-bordeaux-montaigne.fr

IRAMAT-CRP2A (UMR 5060)
Maison de l'Archéologie
Université Bordeaux Montaigne
F-33607 Pessac cedex
France