Extract or Replace Parts of an Object

Operators acting on objects to extract or replace parts.

Usage

# S4 method for GammaSpectrum
[[(x, i)

# S4 method for GammaSpectra
[(x, i, j)

# S4 method for DoseRateModel
[[(x, i)

# S4 method for CalibrationCurve
[[(x, i)

# S4 method for PeakPosition
[[(x, i)

Arguments

x

An object from which to extract element(s) or in which to replace element(s).

i, j

Indices specifying elements to extract or replace. Indices are numeric, integer or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer() (and hence truncated towards zero). Character vectors will be matched to the name of the elements. An empty index (a comma separated blank) indicates that all entries in that dimension are selected.

Value

A subsetted object.

See also

Author

N. Frerebeau